% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.around.decimal.point.r
\name{align.around.decimal.point}
\alias{align.around.decimal.point}
\title{Padding numeric vector around decimal point}
\usage{
\method{align}{around.decimal.point}(x, sep = "\\\\.", len)
}
\arguments{
\item{x}{A vector of numbers.}

\item{sep}{The decimal separator.}

\item{len}{The total character width desired in the formatted field.}
}
\value{
A character vector with members of length \code{len}, with \code{sep} located in the same position
of each member.
}
\description{
Pads a vector of numeric values so that the decimal is at the same ordinal location
}
\examples{
align.around.decimal.point(c(1,100, 0.01))
}
