% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tornado.r
\name{as.tornado.data.frame}
\alias{as.tornado.data.frame}
\title{Coerce to Tornado from data.frame}
\usage{
\method{as.tornado}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{\dots}{ignored}
}
\value{
object of class \code{tornado}
}
\description{
Coerce data.frame to a tornado object.
}
\seealso{
Other tornado: 
\code{\link{as.tornado}()},
\code{\link{plot.tornado}()}
}
\concept{tornado}
