% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tornado.r
\name{as.tornado}
\alias{as.tornado}
\title{Coerce to Tornado}
\usage{
as.tornado(x, ...)
}
\arguments{
\item{x}{object to be coerced}

\item{\dots}{ignored}
}
\value{
object of class \code{tornado}
}
\description{
Coerce something to a tornado object.
}
\seealso{
Other tornado: \code{\link{as.tornado.data.frame}},
  \code{\link{plot.tornado}}
}
\concept{tornado}
