% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asNumeric.r
\name{asNumeric}
\alias{asNumeric}
\title{Convert Character to Numeric with NA Substitution}
\usage{
asNumeric(x, missing.code = c("NA"))
}
\arguments{
\item{x}{A vector of character representations of numbers.}

\item{missing.code}{A string in \code{x} that signifies missing values that should be represented as NA.}
}
\value{
A numeric vector with NA in place of \code{missing.code} .
}
\description{
Converts character to numeric with NA substitution.
}
\examples{
as.numeric(c("3.45","1000","1e6","<BLOQ","NA","<0.100"))
asNumeric(c("3.45","1000","1e6","<BLOQ","NA","<0.100"))
asNumeric(c("3.45","1000","1e6","<BLOQ","NA","<0.100"), missing.code = c("<BLOQ","NA","<0.100"))
}
\seealso{
\code{\link{isMissing}}, \code{\link{isNumeric}}
}
