% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boxcoxf.r
\name{boxcoxf}
\alias{boxcoxf}
\title{Box-Cox Transformation}
\usage{
boxcoxf(shape, distr)
}
\arguments{
\item{shape}{The shape parameter.}

\item{distr}{The distribution parameter.}
}
\value{
Box-Cox tranformation of \code{distr} with \code{shape}.
}
\description{
Apply a Box-Cox Transformation to a vector
}
\examples{
x = rnorm(1000, mean = 0, sd = 0.25)
x.boxcox.pos = boxcoxf(0.5,x)
x.boxcox.neg = boxcoxf(-3.5,x)
par(mfrow=c(1,3))
hist(x)
hist(x.boxcox.pos)
hist(x.boxcox.neg)
}

