% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/catDataFun.r
\name{catDataFun}
\alias{catDataFun}
\title{Summary: length (\%)}
\usage{
catDataFun(y, ndigits.categorical = 1)
}
\arguments{
\item{y}{a vector of values to be summarised.}

\item{ndigits.categorical}{number of digits to round to}
}
\value{
A formatted character vector containing count and percentage of
unique values in \code{y}.
}
\description{
Summary: length (\%)
}
\examples{
catDataFun(data.frame(Sex=rep(c("Female","Male"), times= c(40,60)))$Sex)
}
\seealso{
\code{\link{conDataFun1}},  \code{\link{conDataFun2}},  \code{\link{conDataFun3}},  \code{\link{tabStats}},  \code{\link{tabStats}}
}

