% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conDataFun1.r
\name{conDataFun1}
\alias{conDataFun1}
\title{Summary: mean (SD)}
\usage{
conDataFun1(y, nSignif)
}
\arguments{
\item{y}{numeric vector to summarize}

\item{nSignif}{number of significant digits}
}
\description{
Summary: mean (SD)
}
\note{
This function is primarily used for demographics tables
}
\examples{
x=rnorm(1000)
conDataFun1(x, 3)
}
\seealso{
\code{\link{conDataFun2}},  \code{\link{conDataFun3}},  \code{\link{catDataFun}}, \code{\link{tabStats}}, \code{\link{tabStats}}
}

