% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conDataFun2.r
\name{conDataFun2}
\alias{conDataFun2}
\title{Summary: median (min - max)}
\usage{
conDataFun2(y, nSignif)
}
\arguments{
\item{y}{numeric vector to summarize}

\item{nSignif}{number of significant digits}
}
\description{
Summary: median (min - max)
}
\note{
This function is primarily used for demographics tables
}
\examples{
x=rnorm(1000)
conDataFun2(x, 3)
}
\seealso{
\code{\link{conDataFun1}},   \code{\link{conDataFun3}},  \code{\link{catDataFun}}, \code{\link{tabStats}}, \code{\link{tabSummarize}}
}
