% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conDataFun3.r
\name{conDataFun3}
\alias{conDataFun3}
\title{Summary: (95\%.lower - 95\%.upper)}
\usage{
conDataFun3(y, digits = 3, latex = FALSE, align.dot = FALSE, na.rm = FALSE)
}
\arguments{
\item{y}{numeric vector to summarize}

\item{digits}{number of significant digits}

\item{latex}{passed to \code{\link{formatted.signif}}}

\item{align.dot}{passed to \code{\link{formatted.signif}}}

\item{na.rm}{passed to \code{\link{mean}}}
}
\description{
Summary: (95\%.lower - 95\%.upper)
}
\note{
This function is primarily used for demographics tables
}
\examples{
x = rnorm(1000)
conDataFun3(x, 3)

}
\seealso{
\code{\link{conDataFun1}},  \code{\link{conDataFun2}}, \code{\link{catDataFun}}, \code{\link{tabStats}}, \code{\link{tabSummarize}}
}
