% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.omega.r
\name{convert.omega}
\alias{convert.omega}
\title{Omega Names Conversion}
\usage{
convert.omega(x)
}
\arguments{
\item{x}{named vector (with names being of the format 'OMEGA(x1,x2)').}
}
\value{
reordered named vector.
}
\description{
Convert omega names to match order in covariance and correlation matrix
}
\seealso{
\code{\link{convert.sigma}}
}
