% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.sigma.r
\name{convert.sigma}
\alias{convert.sigma}
\title{Sigma Names Conversion}
\usage{
convert.sigma(x)
}
\arguments{
\item{x}{named vector (with names being of the format 'SIGMA(x1,x2)').}
}
\value{
reordered named vector.
}
\description{
Convert sigma names to match order in covariance and correlation matrix
}
\seealso{
\code{\link{convert.omega}}
}
