% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halflife.r
\name{ConvertPK.iv.Vk.3comp}
\alias{ConvertPK.iv.Vk.3comp}
\title{Calculates coefficients and exponents based on microconstants for 3-compartment IV popPK (post-hoc) parameter estimates}
\usage{
ConvertPK.iv.Vk.3comp(
  ds,
  subjVar = "ID",
  Dose = "DOSE",
  V1 = "V1",
  k10 = "K10",
  k12 = "K12",
  k21 = "K21",
  k13 = "K13",
  k31 = "K31",
  nsig = 3,
  ...
)
}
\arguments{
\item{ds}{# select the dataset}

\item{subjVar}{# select the variable that represents the subject identifier}

\item{Dose}{# select the variables that represents the dose}

\item{V1}{# select the variables that represents the central volume of distribution}

\item{k10}{# select the variables that represents the microconstant k10}

\item{k12}{# select the variables that represents the microconstant k12}

\item{k21}{# select the variables that represents the microconstant k21}

\item{k13}{# select the variables that represents the microconstant k13}

\item{k31}{# select the variables that represents the microconstant k31}

\item{nsig}{# select the number of significant digits}
}
\value{
# returns a data.frame with coefficients, exponents and derived parameters
}
\description{
Calculates coefficients and exponents based on microconstants for 3-compartment IV popPK (post-hoc) parameter estimates
}
\examples{

X <- data

ConvertPK.iv.Vk.3comp(x)

}
\references{
# Dennis Fisher & Steven Shafer NONMEM Workshop - Basic Concepts (2007)

# Upton J Pharmacol Toxicol Methods. 2004 Jan-Feb;49(1):65-8.
}
\author{
# Koen Jolling, Max Lagraauw
}
