% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.process.coveffects.r
\name{covInfo}
\alias{covInfo}
\title{Make a covInfo S3 object}
\usage{
covInfo(
  cov.name,
  cov.vals,
  cov.cat = FALSE,
  cov.center = stats::median(cov.vals),
  cov.min = min(cov.vals),
  cov.max = max(cov.vals),
  cov.label = cov.name,
  cov.breaks = pretty(cov.vals)
)
}
\arguments{
\item{cov.name}{The name of the covariate}

\item{cov.vals}{A vector of values}

\item{cov.cat}{A boolean indicating if the covariate is categorical}

\item{cov.center}{The typical value of the covariate}

\item{cov.min}{The minimum value of the covariate}

\item{cov.max}{The maximum value of the covariate}

\item{cov.label}{The printed name of the covariate (e.g. seen on plot axes)}

\item{cov.breaks}{The points at which simulations and plots would be assessed}
}
\value{
A covInfo object
}
\description{
Constructs a covInfo (covariate information) object from a vector.
}
\note{
This function is used with \code{nm.process.coveffects}
}
\examples{
eqs=make.eqs(CL=THETA1*(1 + (GNDR==0)*THETA7)*(AGE/33.72)^THETA6*exp(ETA1)
            ,V1=THETA2*(1 + (GNDR==0)*THETA9)*(AGE/33.72)^THETA8*exp(ETA2)
            )
eqs$label=list(CL="Clearance (L/h)", V1="Volume of Central Compartment (L)")            

}
\seealso{
\code{\link{print.covInfo}}

Other coveffects: 
\code{\link{eval.eqs}()},
\code{\link{make.eqs}()},
\code{\link{print.covInfo}()}
}
\concept{coveffects}
