% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cut3.r
\name{cut3}
\alias{cut3}
\title{Divide numeric variable into bins}
\usage{
cut3(x, cuts, m = 150, g, levels.mean = FALSE, digits, minmax = TRUE,
  oneval = TRUE, onlycuts = FALSE)
}
\arguments{
\item{x}{numeric vector to be binned}

\item{cuts}{numeric values to cut \code{x} into}

\item{m}{same use as \code{cut2}}

\item{g}{same use as \code{cut2}}

\item{levels.mean}{same value as \code{cut2}}

\item{digits}{same value as \code{cut2}}

\item{minmax}{same value as \code{cut2}}

\item{oneval}{same value as \code{cut2}}

\item{onlycuts}{same value as \code{cut2}}
}
\value{
A numeric vector
}
\description{
Divide numeric variable into bins
}
\note{
This function was based on \code{cut2} from the Hmisc package. Instead of returning factors it returns numeric values representating the mean of the bins
}
\examples{
pkpdData = example.pkpdData()
sunique(cut3(pkpdData$wt, g=10, levels.mean=T))
lunique(cut3(pkpdData$wt, g=10, levels.mean=T))
}

