% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.process.coveffects.r
\name{eval.eqs}
\alias{eval.eqs}
\title{Evaluate a list of equations}
\usage{
eval.eqs(.data, ...)
}
\arguments{
\item{.data}{data.frame, list, or environment}

\item{...}{a list of calls or expressions}
}
\value{
.data is returned with the evaluated expressions added
}
\description{
Evaluates a list of calls in an environment, list, or data.frame
or data.frame.
}
\note{
This function is used with \code{nm.process.coveffects}
}
\examples{
eqs=make.eqs(CL=THETA1*(1 + (GNDR==0)*THETA7)*(AGE/33.72)^THETA6*exp(ETA1)
            ,V1=THETA2*(1 + (GNDR==0)*THETA9)*(AGE/33.72)^THETA8*exp(ETA2)
            )
eqs$label=list(CL="Clearance (L/h)", V1="Volume of Central Compartment (L)")

}
\seealso{
Other coveffects: 
\code{\link{covInfo}()},
\code{\link{make.eqs}()},
\code{\link{print.covInfo}()}
}
\concept{coveffects}
