% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp.examples.r
\name{example.NONMEM.dataset}
\alias{example.NONMEM.dataset}
\title{Examples for qPharmetra style NONMEM data set}
\usage{
example.NONMEM.dataset(ID = 3, TIME = seq(0, 24, 2), DOSE = c(1, 2.5, 10), ...)
}
\arguments{
\item{ID}{number of subjects}

\item{TIME}{sequence of times}

\item{DOSE}{sequence of dose amounts}

\item{\dots}{ignored}
}
\description{
Examples for qPharmetra style NONMEM data set
}
\examples{
library(dplyr)
library(magrittr)
nmData = example.NONMEM.dataset()
nmData \%>\% group_by(DOSE) \%>\% summarise(nID=lunique(ID), nObs = length(DV))
pkpdData = example.pkpdData()
tbl_df(pkpdData)
}
\seealso{
Other qpexamples: 
\code{\link{example.CWRES.plot}()},
\code{\link{example.pkpdData}()},
\code{\link{example.xpose.VPC}()}
}
\concept{qpexamples}
