% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp.examples.r
\name{example.pkpdData}
\alias{example.pkpdData}
\title{Examples for qPharmetra style PKPD data set}
\usage{
example.pkpdData()
}
\description{
Examples for qPharmetra style PKPD data set
}
\examples{
library(dplyr)
pkpdData = example.pkpdData()
tbl_df(pkpdData)
}
\seealso{
Other qpexamples: \code{\link{example.CWRES.plot}},
  \code{\link{example.NONMEM.dataset}},
  \code{\link{example.xpose.VPC}}
}
\concept{qpexamples}
