% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp.examples.r
\name{example.xpose.VPC}
\alias{example.xpose.VPC}
\title{qPharmetra examples for plot, VPC and NONMEM dataset}
\usage{
example.xpose.VPC()
}
\description{
examples for qPharmetra style VPC
}
\examples{
example.xpose.VPC()
}
\seealso{
Other qpexamples: \code{\link{example.CWRES.plot}},
  \code{\link{example.NONMEM.dataset}},
  \code{\link{example.pkpdData}}
}
\concept{qpexamples}
