% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.data.r
\name{expand.data}
\alias{expand.data}
\title{Multiply data.frames}
\usage{
expand.data(data, values, name = "variable.added")
}
\arguments{
\item{data}{input data set}

\item{values}{values of the new column to be added}

\item{name}{name of the column to be added}
}
\description{
Multiply data length(values) times and left join the values as a column
}
\examples{
input.df = data.frame(A = 1:2,B = 2:3,C = 4:5)
input.df
expand.data(input.df, values = c(1,3,7), "Klaas")
}
