% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.character.r
\name{extract.character}
\alias{extract.character}
\title{Extract character information from a string}
\usage{
extract.character(x, collapse = FALSE)
}
\arguments{
\item{x}{A character or numeric string.}

\item{collapse}{A boolean, should the character vector returned by collapsed (see \code{\link[base]{paste}}).}
}
\value{
A character string or vector if the input has multiple character blocks.
}
\description{
Takes character string and parses the character input from it, if available
}
\examples{
extract.character(x="df12Lee31df")
extract.character(x="df1231df", TRUE)
}
\seealso{
\code{\link{extract.number}}
}

