% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.number.r
\name{extract.number}
\alias{extract.number}
\title{Takes character string and parses the numeric input from it, if available.}
\usage{
extract.number(x)
}
\arguments{
\item{x}{A character or numeric string.}
}
\value{
A numeric vector if the input has multiple character blocks.
}
\description{
Takes character string and parses the numeric input from it, if available.
}
\section{Notes}{

If multiple blocks of numeric input exist wthin the input, only the first is parsed
}
\examples{
extract.number("c76")
extract.number("c76c7")
extract.number("BLOQ = 0.144")
}
\seealso{
\code{\link{extract.character}}
}

