% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.varcov.r
\name{extract.varcov}
\alias{extract.varcov}
\title{Get the covariance matrix from an already parsed .cov output file.}
\usage{
extract.varcov(text)
}
\arguments{
\item{text}{A character or numeric string containing the .cov output.}
}
\value{
A named list. The sole item in the list should have a name that is the estimation method 
used (e.g. "First Order Conditional Estimation with Interaction").  That named item will be a 
data frame containing the covariance matrix.
}
\description{
Get the covariance matrix from an already parsed .cov output file.
}
\section{Notes}{

function not to be called alone - is called by qP function nm.covmat.extract inside get.covmat
}

\examples{
\dontrun{
nm.unzip(
  path = file.path(getOption("qpExampleDir"),"run11"),
  run = "run11",extension = ".cov"
)
covText = scan(
  file = file.path(getOption("qpExampleDir"),"run11","run11.cov"), 
  what = "character",sep = "\n", quiet = T
)

## the heart of the function: parse space delimited text into a matrix
cov = extract.varcov(covText[-1])

## warning: this matrix is LARGE. View only start and end of output
head(cov)
tail(cov)

file.remove(file.path(getOption("qpExampleDir"),"run11","run11.cov")) ## clean things up
}
}
\seealso{
\code{\link{nm.covmat.extract}}, \code{\link{get.covmat}}
}
