% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formalize.R
\name{formalize.data.frame}
\alias{formalize.data.frame}
\title{Formalize Data Frame}
\usage{
\method{formalize}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{passed arguments}
}
\value{
formalized data.frame
}
\description{
Formalizes data.frame.
Replaces column names with labels, where present.
Stores column name as 'alias' attribute.
}
\examples{
x <- data.frame(x = 1:10, y = 1:10, z = 1:10)
attr(x$x, 'label') <- 'Independent Value'
attr(x$y, 'label') <- 'Dependent Value'
x
formalize(x)
}
\seealso{
Other formalize: 
\code{\link{formalize}()}
}
\concept{formalize}
