% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formalize.R
\name{formalize}
\alias{formalize}
\title{Formalize Something}
\usage{
formalize(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Formalizes something.
Generic, with method \code{\link{formalize.data.frame}}.
}
\seealso{
Other formalize: 
\code{\link{formalize.data.frame}()}
}
\concept{formalize}
\keyword{internal}
