% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format.demoTable.r
\name{format.demoTable}
\alias{format.demoTable}
\title{Format a tabulated summary for LaTeX}
\usage{
\method{format}{demoTable}(demoTable, formula)
}
\arguments{
\item{demoTable}{output from \code{tabSummarize}}

\item{formula}{the same formula as used to create the \code{tabSummarize} output}
}
\value{
Table with summarized data formatted for inclusion in LaTeX
}
\description{
Format a tabulated summary for LaTeX
}
\note{
This function is primarily used for demographics tables
}
\examples{
options(width = 150)
pkpdData = example.pkpdData()
ok = duplicated(pkpdData$id) == F
myFormula =  dose ~ race + wt + bmi + sex
my.formatted.demoTable = format.demoTable(
 tabSummarize(formula = myFormula, data = pkpdData[ok, ], nSignif = 3)
  , formula = myFormula
)
print(my.formatted.demoTable
  , sanitize.text.function = identity
  , booktabs = T
)
}
\seealso{
\code{\link{tabStats}}, \code{\link{tabSummarize}}
}

