% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatted.signif.r
\name{formatted.signif}
\alias{formatted.signif}
\title{Forced length significant digits}
\usage{
formatted.signif(x, digits = 3, latex = F, align.dot = F)
}
\arguments{
\item{x}{a vector of values to be processed into significant digits format}

\item{digits}{single value specifying the number of significant digits}
}
\value{
A formatted character vector containing \code{x} in with significant \code{digits} format
}
\description{
Forced length significant digits
}
\examples{
formatted.signif(c(5,4.99, 4.99999,5.0000001,5001),4)

numvec = c(5,4.99, 4.99999,5.0000001,5001, 0.00005,101,39.9)
print(
   xtable(
      data.frame(original=numvec
                 ,latex.TRUE= formatted.signif(numvec,3, T,F)
                 ,align.TRUE= formatted.signif(numvec,3, F,T)
                 ,all.TRUE = formatted.signif(numvec,3, T,T)
      )
   )
   , booktabs=T
   , sanitize.text.function = identity
)
# now take this to a LaTeX compiler!
}
\seealso{
\code{\link{signif}}
}

