% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/full.names.r
\name{full.names}
\alias{full.names}
\title{swaps shorthand / abbreviated column names in data frame for full names (units) for labelling purposes}
\usage{
full.names(x)
}
\arguments{
\item{x}{Shorthand / Abbreviate column names}
}
\value{
Column names written out in full
}
\description{
swaps shorthand / abbreviated column names in data frame for full names (units) for labelling purposes
}
\examples{
full.names(c("wt","race","age"))
}

