% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxs.r
\name{fxs}
\alias{fxs}
\title{Fixed character length of a String}
\usage{
fxs(x, pattern = "0000")
}
\arguments{
\item{x}{string to format}

\item{pattern}{format to adhere to}
}
\value{
x reformatted according to pattern
}
\description{
Format string of fixed character length
}
\examples{
fxs(c(0,1,34, 50000), pattern="    ")
fxs(c(0,1,34, 50000), "000000")
}
