% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.covmat.r
\name{get.covmat}
\alias{get.covmat}
\title{Get the covariance matrix}
\usage{
get.covmat(run, path = paste(getwd(), "NONMEM", sep = "/"),
  ext = "cov")
}
\arguments{
\item{run}{run rootname (e.g. run1)}

\item{path}{directory where run output resides}

\item{ext}{extension of the covariance}
}
\value{
list with covariance matrices for each estimation method
}
\description{
Get the covariance matrix from a completed NONMEM run. If the the run has been zipped, the file will be extracted, then parsed and thereafter the file (not the zipped file) will be discarded.
}
\examples{
get.covmat("example2", path = getOption("qpExampleDir"))
}
\seealso{
\code{\link{nm.covmat.extract}}, \code{\link{extract.varcov}}
}
