% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.description.r
\name{get.description}
\alias{get.description}
\title{Scan the runrecord tag for 'Description'}
\usage{
get.description(run, path = getOption("nmDir"), file.ext = ".lst")
}
\arguments{
\item{run}{run rootname (e.g. run1)}

\item{path}{directory where NONMEM output file resides}

\item{file.ext}{extension of the .ext file}
}
\value{
The model's description tag as a character string
}
\description{
Scan the runrecord tag for 'Description'. This makes it easy to scan a few models to find some particular model
}
\examples{
get.description(run = "example1", path = getOption("qpExampleDir"))
}
\seealso{
\code{\link{get.nm.version}}, \code{\link{get.ofv}}, \code{\link{read.out}}
}

