% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.diag.names.r
\name{get.diag.names}
\alias{get.diag.names}
\title{Find diagonal OMEGA names}
\usage{
get.diag.names(names, sep = ",")
}
\arguments{
\item{names}{character vector with OMEGA and SIGMA names}

\item{sep}{separator - defaults to comma (e.g. OMEGA(1,1))}
}
\value{
named vector of logicals; TRUE for giagonal, FALSE for off-diagonal elements
}
\description{
Retrieve which names in a vector of NONMEM parameter names are the diagonal ones. Thus, slect which ones are (1,1),, (2,2), etc... instead of (1,2), (3,4), etc...  The function is used internally by functions post-processing NONMEM output
}
\examples{
   
covMatrix = get.covmat(path = getOption("qpExampleDir"), run = "example2")
get.diag.names(names = sub(" +", "", trimSpace(names(covMatrix[[1]]))))
}
