% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.dOFV.r
\name{get.dOFV}
\alias{get.dOFV}
\title{Create brief NONMEM run record}
\usage{
get.dOFV(runs = c("run1", "run2"), path = getOption("nmDir"),
  conf.level = 0.95, file.ext = ".lst")
}
\arguments{
\item{runs}{character vector of existing NONMEM runs}

\item{path}{directory where NONMEM output resides}

\item{conf.level}{confidence level - currently not used}

\item{file.ext}{file extension of the NONMEM output file (defaults to .lst)}
}
\value{
Placeholder for return description
}
\description{
rather than calling PsN's runrecord (the full blown parser) get.dOFV create a brief runrecord on the fly. It might be used for interim reports and presentations but ideally formal reporting uses
}
\examples{
get.dOFV(runs = c("example1","example2"), path = getOption("qpExampleDir"))
# note this example is for two non-nested models. It has not other value to to illustrate the tool
}
\seealso{
\code{\link{get.ofv}} which is called inside \code{get.dOFV}, \code{\link{read.out}}
}
