% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.eigen.r
\name{get.eigen}
\alias{get.eigen}
\title{NONMEM run Eigen values}
\usage{
get.eigen(run, path = getOption("nmDir"), file.ext = ".lst")
}
\arguments{
\item{run}{run rootname (e.g. \code{run1})}

\item{path}{directory where \code{run} resides}

\item{file.ext}{extensionof the NONMEM output file (.lst)}
}
\value{
list with named vectors with Eigen values for each estimation method
}
\description{
Get the Eigen values from a NONMEM run. Requires the user to have run NONMEM with $COV ... PRINT=E option.
}
\examples{
get.eigen(run = "example1", path = getOption("qpExampleDir"))
get.eigen(run = "example2", path = getOption("qpExampleDir"))
}
\seealso{
\code{\link{get.nm.version}}, \code{\link{get.shrinkage}}, \code{\link{get.ofv}}, \code{\link{read.out}}
}

