% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.multiple.xpose.tables.r
\name{get.multiple.xpose.tables}
\alias{get.multiple.xpose.tables}
\title{Get xpose tables of multiple runs}
\usage{
get.multiple.xpose.tables(runs, path = getOption("nmDir"),
  carryAlong = c("DV", "DOSE", "TIME", "EVID"))
}
\arguments{
\item{runs}{character vector of run roon names (e.g. run1)}

\item{path}{directory where run directories of runs reside}

\item{carryAlong}{character vector of variables to get in addition to ID, PRED, IPRED, CWRES, DV. Argument carryAlong defaults to DOSE TIME and EVID. Use in case output tables are not consistent across runs being imported}
}
\value{
data.fame the xpose tables for runs and an additional column 'model' containinng the rootnames of runs for stratification in post-processing
}
\description{
Imports output tables of multiple runs and merges them into one data.frame
}
\note{
use argument 'carryAlong' to ask for more data items to be included in the result
}
\examples{
library(Hmisc)
library(lattice)
 test = get.multiple.xpose.tables(runs = c("example1","example2")
 , path = getOption("qpExampleDir"), carryAlong = c("CONC","TIME","EVID"))
names(test)
head(subset(test,model == "example1"))
head(subset(test,model == "example2"))
xyplot(CWRES ~ value | variable * model
       , data = reshape2::melt(test, measure.vars = Cs(TIME,PRED))
       , panel = panel.cwres
       , scales = list(x = list(relation = "free"))
)
}
\seealso{
\code{\link{get.xpose.tables}}
}
