% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.nm.version.r
\name{get.nm.version}
\alias{get.nm.version}
\title{NONMEM version used}
\usage{
get.nm.version(run, path = getOption("nmDir"), file.ext = ".lst")
}
\arguments{
\item{run}{run rootname (e.g. run1)}

\item{path}{= directory where NONMEM run resides}

\item{file.ext}{NONMEM output file extension (defaults to .ext)}
}
\value{
character string with NONMEM version used
}
\description{
Get NONMEM version (7.3.0 / 7.2.x etc )
}
\examples{
get.nm.version(run = "example1", path = getOption("qpExampleDir"))
}
\seealso{
\code{\link{read.out}}
}
