% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ofv.r
\name{get.ofv}
\alias{get.ofv}
\title{Objective Function Value parser}
\usage{
get.ofv(run, path = getOption("nmDir"), file.ext = ".lst")
}
\arguments{
\item{run}{run rootname (e.g. \code{run1})}

\item{path}{directory where \code{run} resides}

\item{file.ext}{file extension of the NONMEM output file (defaults to .lst)}
}
\value{
named numeric value representing the OFV. The name is NONMEM's estimated-method-specific name for the OFV
}
\description{
Pull the Objective Function Value (OFV) from the NONMEM output file
}
\examples{
get.ofv("example2", path = getOption("qpExampleDir"))
sapply(c("example1","example2"), get.ofv, path = getOption("qpExampleDir"))
}
\seealso{
\code{\link{get.dOFV}} which calls \code{get.ofv} over multiple runs for a brief runrecord, \code{\link{read.out}}
}
