% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.omega.r
\name{get.omega}
\alias{get.omega}
\title{Summary: OMEGA estimates}
\usage{
get.omega(run, path = getOption("nmDir"), ...)
}
\arguments{
\item{run}{run rootname (e.g. 'run1')}

\item{path}{directory where rootname.ext resides}

\item{...}{additional arguments passed to \code{nm.params.table}.}
}
\value{
data frame with OMEGA estimates, coefficient of variation, standard error, and estimated/fixed information.
}
\description{
Subset parameter estimate table by OMEGAs
}
\examples{
get.omega("example2",  path = getOption("qpExampleDir"), fixed.text = "(fixed to 0)")
}
\seealso{
\code{\link{nm.params.table}}, \code{\link{get.theta}}, \code{\link{get.sigma}}
}

