% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.sigma.r
\name{get.sigma}
\alias{get.sigma}
\title{Summary: SIGMA estimates}
\usage{
get.sigma(run, path = getOption("nmDir"), ...)
}
\arguments{
\item{run}{run rootname (e.g. 'run1')}

\item{path}{directory where rootname.ext resides}

\item{...}{additional arguments passed to \code{nm.params.table}.}
}
\value{
data frame with SIGMA estimates, coefficient of variation, standard error, and estimated/fixed information.
}
\description{
Subset parameter estimate table by SIGMAs
}
\examples{
get.sigma("example2",  path = getOption("qpExampleDir"), fixed.text = "(fixed to 0)")
}
\seealso{
\code{\link{nm.params.table}}, \code{\link{get.theta}}, \code{\link{get.omega}}
}

