% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npar.R
\name{get.theta.npar}
\alias{get.theta.npar}
\title{Calculate number of Thetas}
\usage{
get.theta.npar(model, path = getOption("nmDir"), prefix = "run", ...)
}
\arguments{
\item{model}{select the model run number (e.g. 1)}

\item{path}{select the directory where the model file resides}

\item{prefix}{text to precede run number in model name (default: 'run')}

\item{...}{additional arguments passed to \code{\link{get.theta}}}
}
\value{
returns a vector with the number of estimated THETA parameters
}
\description{
Calculates the number of THETA parameters 
that are estimated (and not fixed) 
in the NONMEM model file.
}
\examples{
\dontrun{\donttest{
models = c(101,102,104:109,111,115, 141:145, 147:149, 151, 153:156)
get.theta.npar(models)
}}


}
\author{
Koen Jolling
}
