% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.xpose.tables.r
\name{get.xpose.tables}
\alias{get.xpose.tables}
\title{Get NONMEM output tables into one data frame}
\usage{
get.xpose.tables(run, path = getOption("nmDir"))
}
\arguments{
\item{run}{character vector of run roon names (e.g. run1)}

\item{path}{directory where run directories of runs reside}
}
\value{
data frame with all (unique) data items exported in Xpose tables sdtab, patab, catab and cotab
}
\description{
Get NONMEM output tables in XPOSE format into one data frame. Requires all files to be created without the FIRSTONLY option in $TABLE. Also, ensure ONEHEADER option is consistently used in all $TABLE statements
}
\examples{
out.1 = get.xpose.tables(path = getOption("qpExampleDir"), run = "example2")
head(out.1)
}
\seealso{
\code{\link{get.multiple.xpose.tables}}
}
