% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.bootstrap.r
\name{histogram.bootstrap}
\alias{histogram.bootstrap}
\title{Bootstrap Histograms}
\usage{
\method{histogram}{bootstrap}(bootstrap,
  filename.prefix = "bootstrapPage", path, incl.ids = NULL,
  min.failed = FALSE, cov.failed = FALSE, cov.warnings = FALSE,
  boundary = FALSE, showoriginal = TRUE, showmean = TRUE,
  obsCentralCol = "blue", bootCentralCol = "red",
  outerCol = "darkslategray", showmedian = FALSE, show95CI = TRUE,
  conf.int = 0.95, showquart = FALSE, histCol = "#F2F2F2",
  densityCol = "red", excl.id = c(), excl.columns = c("model",
  "minimization.successful", "covariance.step.successful",
  "covariance.step.warnings", "estimate.near.boundary"), ...)
}
\arguments{
\item{bootstrap}{Output generated by read.bootstrap}

\item{filename.prefix}{Prefix for filenames of PDFs created,}

\item{path}{Directory where graph files should appear (default reads from output)}

\item{incl.ids}{Include individuals in the plot? Defaults to NULL}

\item{min.failed}{Logical (defaults to FALSE) do we want to omit minimization failed runs?}

\item{cov.failed}{Logical (defaults to FALSE) do we want to omit covariance failed runs?}

\item{cov.warnings}{Logical (defaults to FALSE) do we want to omit covariance failed runs?}

\item{boundary}{Logical (defaults to FALSE) do we want to omit boundary runs?}

\item{showoriginal}{Logical (defaults to TRUE) to show line for original estimate}

\item{showmean}{Logical (defaults to TRUE) to show line for mean}

\item{obsCentralCol}{Color of observed central tendency}

\item{bootCentralCol}{Color of bootstrapped central tendency}

\item{outerCol}{Color of outer region polygon}

\item{showmedian}{Logical (defaults to FALSE) to show line for median}

\item{show95CI}{Logical (defaults to TRUE) show line for 95\% confidence interval (percentile)}

\item{conf.int}{Confidence level for bootstrap sample outer bands}

\item{showquart}{Logical (defaults to FALSE) to show line for quartiles}

\item{histCol}{Histogram bar color}

\item{densityCol}{Color of polygon}

\item{excl.id}{Exclude samples that have this individual}

\item{excl.columns}{What elements should not be shown?}

\item{...}{Optional arguments passed on to hist}
}
\value{
Histograms of bootstrapped parameter estimates
}
\description{
Create histograms of PsN boostrap output read with read.bootstrap
}
\examples{
myBoot = read.bootstrap(path = getOption("qpExampleDir"),
filename = "bootstrap/raw_results_bs4011.csv",
structure.filename = "bootstrap/raw_results_structure")
temp.dir = tempdir()
cat(temp.dir)
histogram.bootstrap(myBoot, path = temp.dir, filename.prefix = "Test") 
## now take a look there
}
