% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.blanks.r
\name{insert.blanks}
\alias{insert.blanks}
\title{Inserting blank lines in a table}
\usage{
insert.blanks(tab, lines)
}
\arguments{
\item{tab}{the table to insert the line into}

\item{lines}{vector of numeric row positions in \code{tab} to insert a line}
}
\value{
a table with lines inserted
}
\description{
Create a blank line in a table typically used for legibility improvement when reporting a table.
}
\examples{
pkpdData = example.pkpdData()
tmp = pkpdData[1:3,]
tmp
insert.blank.line(tmp, 2)
}
