% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qP.colors.r
\docType{data}
\name{ketchup}
\alias{ketchup}
\title{ketchup}
\format{An object of class \code{character} of length 1.}
\usage{
ketchup
}
\description{
ketchup
}
\seealso{
Other qpcolors: 
\code{\link{PI.ci.down.arcol}},
\code{\link{PI.ci.med.arcol}},
\code{\link{PI.real.down.col}},
\code{\link{PI.real.med.col}},
\code{\link{PI.real.up.col}},
\code{\link{abline.color.box}},
\code{\link{abline.color}},
\code{\link{amber}},
\code{\link{apple}},
\code{\link{blue}},
\code{\link{box.plot.color}},
\code{\link{brown}},
\code{\link{cobalt}},
\code{\link{crimson}},
\code{\link{cyan}},
\code{\link{emerald}},
\code{\link{gray}},
\code{\link{green}},
\code{\link{histogram.color}},
\code{\link{indigo}},
\code{\link{ipred.color}},
\code{\link{lime}},
\code{\link{lin.fit.col}},
\code{\link{magenta}},
\code{\link{mauve}},
\code{\link{obs.color}},
\code{\link{olive}},
\code{\link{orange}},
\code{\link{pink}},
\code{\link{pred.color}},
\code{\link{purple}},
\code{\link{qp.blue}},
\code{\link{qp.colors}},
\code{\link{qp.green}},
\code{\link{red}},
\code{\link{smooth.ci.color}},
\code{\link{smooth.color}},
\code{\link{steel}},
\code{\link{taupe}},
\code{\link{teal}},
\code{\link{violet}},
\code{\link{yellow}}
}
\concept{qpcolors}
\keyword{datasets}
