% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.update.glossares.r
\name{latex.update.glosssaries}
\alias{latex.update.glosssaries}
\title{Update Glossaries}
\usage{
latex.update.glosssaries(
  dir = ".",
  base = "Report",
  command = "makeindex",
  s = paste0("-s ", base, ".ist"),
  t = paste0("-t ", base, ".glg"),
  o = paste0("-o ", base, ".gls"),
  p = paste0("", base, ".glo"),
  args = paste(s, t, o, p),
  call = paste(command, args),
  quiet = TRUE,
  ...
)
}
\arguments{
\item{dir}{path to directory containing the report}

\item{base}{basename for the report}

\item{command}{defaults to 'makeindex'}

\item{s}{default -s argument (*.ist)}

\item{t}{default -t argument (*.glg)}

\item{o}{default -o argument (*.gls)}

\item{p}{default -p argument (*.glo)}

\item{args}{arguments to command; overrides s, t, o, p}

\item{call}{text of system call; overrides command and args}

\item{quiet}{suppress messages}

\item{...}{ignored}
}
\value{
Used for side effects.
}
\description{
Updates glossaries in latex-based report. Changes to the specified directory,
and should change back even if an error occurs.
}
