% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{LIBRARY}
\alias{LIBRARY}
\title{Load a Specific Package Version}
\usage{
LIBRARY(
  package,
  version = 0,
  help,
  pos = 2,
  lib.loc = NULL,
  character.only = TRUE,
  logical.return = FALSE,
  warn.conflicts,
  quietly = FALSE,
  verbose = getOption("verbose"),
  mask.ok,
  exclude,
  include.only,
  attach.required = missing(include.only)
)
}
\arguments{
\item{package}{package name, possibly unquoted}

\item{version}{minimum allowed version}

\item{help}{passed to \code{\link{library}}}

\item{pos}{passed to \code{\link{library}}}

\item{lib.loc}{passed to \code{\link{library}}}

\item{character.only}{passed to \code{\link{library}}}

\item{logical.return}{passed to \code{\link{library}}}

\item{warn.conflicts}{passed to \code{\link{library}}}

\item{quietly}{passed to \code{\link{library}}}

\item{verbose}{passed to \code{\link{library}}}

\item{mask.ok}{passed to \code{\link{library}}}

\item{exclude}{passed to \code{\link{library}}}

\item{include.only}{passed to \code{\link{library}}}

\item{attach.required}{passed to \code{\link{library}}}
}
\description{
Loads a specific package version.
}
\examples{
LIBRARY(knitr, '0.9.0')
}
\keyword{internal}
