% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logit.inv.r
\name{logit.inv}
\alias{logit.inv}
\title{Logit Back-transformation}
\usage{
logit.inv(x, min = 0, max = 1)
}
\arguments{
\item{x}{any value}

\item{min}{lower boundary of backtransformed value}

\item{max}{upper boundary of backtransformed value}
}
\value{
logit(x) backtransformed to <0,1>
}
\description{
Backtransforms logit transformed data to the <0,1> scale
}
\examples{
myVector = runif(n = 1000, min=0.001,max = 0.999)
myVector.logit = logit(myVector)
par(mfrow = c(1,3))
hist(myVector)
hist(myVector.logit)
hist(logit.inv(myVector.logit))
}
\seealso{
\code{\link{logit}}
}

