% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logit.r
\name{logit}
\alias{logit}
\title{Logit transformation}
\usage{
logit(x, min = 0, max = 1)
}
\arguments{
\item{x}{any value}

\item{min}{lower boundary of backtransformed value}

\item{max}{upper boundary of backtransformed value}
}
\value{
Logit-transformed x
}
\description{
Logit-transforms any value between 0 and 1
}
\examples{
myVector = runif(n = 1000, min=0.001,max = 0.999)
myVector.logit = logit(myVector)
par(mfrow = c(1,3))
hist(myVector)
hist(myVector.logit)
hist(logit.inv(myVector.logit))
}
\seealso{
\code{\link{logit.inv}}
}

