% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qP.colors.r
\docType{data}
\name{loi.color}
\alias{loi.color}
\title{loi.color (Autograph)}
\format{An object of class \code{character} of length 1.}
\usage{
loi.color
}
\description{
loi.color (Autograph)
}
\seealso{
Other qpcolors (Autograph): \code{\link{PI.ci.up.arcol}}
}
\concept{qpcolors (Autograph)}
\keyword{datasets}
