% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lunique.r
\name{lunique}
\alias{lunique}
\title{Number of unique values}
\usage{
lunique(x)
}
\arguments{
\item{x}{values to be uniqued}
}
\value{
Number of unique values in vector \code{x}
}
\description{
Derive the length (number) of unique values in a vector
}
\examples{
lunique(1:10)
lunique(c(1,1,2,2,3,3,4,4))
}
\seealso{
\code{\link{sunique}}
}

