% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.process.coveffects.r
\name{make.eqs}
\alias{make.eqs}
\title{Make a list of equations}
\usage{
make.eqs(...)
}
\arguments{
\item{...}{a series of named formulas}
}
\value{
a list of call objects that can be used to evaluate the input equations 
in a given environment.
}
\description{
Makes a list of equations.
input formulas.
}
\note{
This function is used with \code{nm.process.coveffects}
}
\examples{
eqs=make.eqs(CL=THETA1*(1 + (GNDR==0)*THETA7)*(AGE/33.72)^THETA6*exp(ETA1)
            ,V1=THETA2*(1 + (GNDR==0)*THETA9)*(AGE/33.72)^THETA8*exp(ETA2)
            )
eqs$label=list(CL="Clearance (L/h)", V1="Volume of Central Compartment (L)")            

}
\seealso{
Other coveffects: \code{\link{covInfo}},
  \code{\link{eval.eqs}}, \code{\link{print.covInfo}}
}
\concept{coveffects}
