% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.symmetric.r
\name{make.symmetric}
\alias{make.symmetric}
\title{Creat symmetrical matrix}
\usage{
make.symmetric(values, triangle = "upper")
}
\arguments{
\item{values}{vector of length cumsum(ncol(matrix))}

\item{triangle}{defaulting to 'upper', can also take the value 'lower'}
}
\value{
A matrix of dimension ncol x ncol
}
\description{
Create symmetrical matrix of lower or upper traingle of a matrix
}
\examples{
  make.symmetric(c(0.12,
                   0.02,1.02,
                   0.23,-0.3,0.23))
}
\seealso{
\code{\link{xml.Extract.NM.Matrix}}
}
