% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.process.coveffects.r
\name{makeCovInfo}
\alias{makeCovInfo}
\title{Make a list of covInfo objects from a data frame.}
\usage{
makeCovInfo(df, cnames = names(df), cat.covs = "")
}
\arguments{
\item{df}{A data frame}

\item{cnames}{A vector of column names to create covInfo objects for}

\item{cat.covs}{A vector of categorical column names.  Must be a subset of cnames}
}
\value{
A list of covInfo objects
}
\description{
Takes a data frame and creates covInfo objects based on its columns.
}
\note{
This function is used with \code{nm.process.coveffects}
}
