% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.create.control.r
\name{makeRandom}
\alias{makeRandom}
\title{Create a control stream for simulation}
\usage{
makeRandom(sw = "omega", ctl, ranEf)
}
\arguments{
\item{sw}{either "omega" or "sigma" for IIV or residual random effects}

\item{ctl}{character vector containing the current version of the NONMEM control stream}

\item{ranEf}{vector with the estimated random effects. These are the estimates to be placed in the new control stream}
}
\value{
A control stream ready for execution written to file
}
\description{
Parse random effects and construct new random effects statements
}
\note{
This function is only used by \code{nm.create.control}.
}
\seealso{
\code{\link{nm.create.control}}
}
