% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meta.map.r
\name{meta.map}
\alias{meta.map}
\title{Draw meta-map}
\usage{
meta.map(meta.ds, compNm, trialNm, refNm, ref.col = blue[2],
  comp.col = green[3], dist = 1.35)
}
\arguments{
\item{meta.ds}{meta analysis dataset}

\item{compNm}{name of variable holding the comparator names}

\item{trialNm}{name of variable holding unique trial identifier}

\item{refNm}{Reference comparator. Must be a member of meta.ds[,compNm]}

\item{ref.col}{color of reference node}

\item{comp.col}{color of comparator nodes}

\item{dist}{distance for text position}
}
\value{
A graph mapping the dimensions of the meta-analysis dataset
}
\description{
Draw a mapping of a meta-analysis dataset
}
\examples{
test = c("Ref", paste("comp", 1:5, sep=""))
test.ds = data.frame(drg = rep(test, times=seq(12,2,-2)),
                     ref.id = c(rep(1:4, 10),1,1))

names(test.ds)
meta.map(test.ds, compNm="drg", trialNm="ref.id", refNm="Ref", dist=1.4)
}
\seealso{
\code{\link{make.map}}
}

