% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro2macro.r
\name{micro2macro}
\alias{micro2macro}
\title{Convert Micro Parameters to Macro Parameters}
\usage{
micro2macro(..., drop = TRUE)
}
\arguments{
\item{...}{named parameters V1, k10, and (conditionally) k12, k21, k13, k31}

\item{drop}{exclude micro parameters from the result}
}
\value{
a list
}
\description{
Conditionally converts a list of micro parameters to macro parameters
for up to 3 compartments. Recognizes V1, k10, k12, k21, k13, k31.
Returns CL, V1, and (conditionally) Q2, V2, Q3, and V3.
}
\examples{

library(pmxTools)
list(
    V1  =  c(2, 3),
    k10 = c(1.2, 1.3), 
    k12 = c(12,   11),
    k21 = c(2.3, 2.4),
    k13 = c(3.2, 3.0),
    k31 = c(0.04, 0.05)
  ) \%>\% 
  do.call(micro2macro, .) \%>\%
  do.call(calc_derived, .) \%>\% 
  subset(lengths(.) != 0) \%>\%
  data.frame
  
}
