% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss.col.r
\name{miss.col}
\alias{miss.col}
\title{Detect differences between two data.frames}
\usage{
miss.col(df.1, df.2)
}
\arguments{
\item{df.1}{first data.frame}

\item{df.2}{second data.frame}
}
\value{
Placeholder for return description
}
\description{
given two data frames, which if any column in the first is not found in the second
}
\examples{
my.df.1 = data.frame(a = 1:3,b = 4:6,c = 7:9)
my.df.2 = data.frame(a = 1:3,b = 4:6)

miss.col(my.df.1,my.df.2)
#"c"
miss.col(my.df.2,my.df.1)
#character(0)
}
