% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPNG.r
\name{myPNG}
\alias{myPNG}
\title{General PNG grpah creator}
\usage{
myPNG(
  filename = "C:/Test/test.png",
  w = 8,
  h = 6,
  units = "in",
  psize = 10,
  bg = "transparent",
  res = 600,
  ...
)
}
\arguments{
\item{filename}{filename including path. To follow qP workflow standards should read the internal relative reference directory "../WorkAre/output/graphs"}

\item{w}{width in \code{units}}

\item{h}{height in \code{units}}

\item{units}{units ('in', 'cm', or 'pt')}

\item{psize}{pointsize in \code{units}}

\item{bg}{background color}

\item{res}{resolution of PNG graph. High values create top quality graph at the expense of large file size.}

\item{...}{any additional parameters to be passed on the \code{png()}}
}
\value{
A graph written to file in .png format.
}
\description{
A general wrapper function for a call to png(), which opens a plot device (i.e. enables you to "print" a graphic to a file). Note that the plot must be generated and then \code{dev.off()} must be called to close the plot device (i.e. close the output graphic file)
}
