% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme.extract.r
\name{nlme.extract}
\alias{nlme.extract}
\title{Extract information from nlme object}
\usage{
\method{nlme}{extract}(
  obj,
  method = "components",
  newdata = NULL,
  sample = FALSE,
  nSignif = 3,
  nSignif.CV = 0
)
}
\arguments{
\item{obj}{an nlme object}

\item{method}{can be either components (extracting function) or samples (for VPC)}

\item{newdata}{a new data set with at the minimum the data items (independent, dependent variables, covariates, and grouping variables) required to build the nlme object.}

\item{sample}{logical (default is FALSE) indicating whether to sample or not}

\item{nSignif}{number of significant digits for parameter estimate table}

\item{nSignif.CV}{number of significant digits for coefficient of variation in the parameter estimate table}
}
\value{
list with extracted nlme object information, model summary, model parameter table, or, when method = "samples", a matrix with predictions for data.frame \code{getData(obj)} or \code{newdata}.
}
\description{
Given an nlme object, extract all available information for further processing. In addition parse the information into a ready-to-use parameter estimate table. Furthermore, when method is set to samples, it will create samples to be used for a VPC (see \code{nlme.vpc}).
}
\examples{
fm1 <- nlme::nlme(height ~ SSasymp(age, Asym, R0, lrc),
data = Loblolly,
fixed = Asym + R0 + lrc ~ 1,
random = Asym ~ 1,
start = c(Asym = 103, R0 = -8.5, lrc = -3.3))
fm1.out = nlme.extract(fm1)
fm1.out
}
\seealso{
\code{\link{nlme.run}}, \code{\link{nlme.diag}}, \code{\link{nlme.vpc}},  \code{\link{nlme.simPars}}, \code{\link{nlme.getFixPars}}, \code{\link{nlme.getRanPars}}
}
