% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlme.getFixPars.r
\name{nlme.getFixPars}
\alias{nlme.getFixPars}
\title{NLME fixed effects and covariance matrix}
\usage{
\method{nlme}{getFixPars}(obj)
}
\arguments{
\item{obj}{an nlme object}
}
\value{
list with fixed effect point estimates and variance-covariance matrix
}
\description{
extracts fixed effect point estimates and variance-covariance matrix of an nlme object
}
\examples{
library(nlme)
fm1 <- nlme(height ~ SSasymp(age, Asym, R0, lrc),
  data = Loblolly,
  fixed = Asym + R0 + lrc ~ 1,
  random = Asym ~ 1,
  start = c(Asym = 103, R0 = -8.5, lrc = -3.3)
  )
nlme.getFixPars(fm1)
}
\seealso{
\code{\link{nlme.getRanPars}}, \code{\link{nlme.simPars}}, \code{\link{nlme.extract}}
}

