% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme.getRanPars.r
\name{nlme.getRanPars}
\alias{nlme.getRanPars}
\title{NLME random effects and covariance matrix}
\usage{
\method{nlme}{getRanPars}(object)
}
\arguments{
\item{object}{an nlme object}
}
\value{
list with fixed effect point estimates and variance-covariance matrix
}
\description{
Extracts random effect point estimate and variance-covariance matrix of an nlme object
}
\examples{
library(nlme)
fm1 <- nlme(height ~ SSasymp(age, Asym, R0, lrc),
  data = Loblolly,
  fixed = Asym + R0 + lrc ~ 1,
  random = Asym ~ 1,
  start = c(Asym = 103, R0 = -8.5, lrc = -3.3)
  )
nlme.getFixPars(fm1)
}
\seealso{
\code{\link{nlme.getFixPars}}, \code{\link{nlme.simPars}}, \code{\link{nlme.extract}}
}
