% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlme.simPars.r
\name{nlme.simPars}
\alias{nlme.simPars}
\title{Simulate from an nlme object}
\usage{
\method{nlme}{simPars}(paramList, N = 1)
}
\arguments{
\item{paramList}{output from nlme.FixPars or nlme.RanPars}

\item{N}{sample size of the requested predictions}
}
\value{
Matrix with parameter prediction
}
\description{
Draw samples from multivariate normal parameter distribution from an nlme object
}
\examples{
library(nlme)
fm1 <- nlme(height ~ SSasymp(age, Asym, R0, lrc),
  data = Loblolly,
  fixed = Asym + R0 + lrc ~ 1,
  random = Asym ~ 1,
  start = c(Asym = 103, R0 = -8.5, lrc = -3.3)
  )
nlme.getFixPars(fm1)
nlme.getRanPars(fm1)
nlme.simPars(nlme.getFixPars(fm1), N = 10)
}
\seealso{
\code{\link{nlme.getFixPars}}, \code{\link{nlme.getRanPars}}, \code{\link{nlme.extract}}
}

