% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme.vpcplot.r
\name{nlme.vpcplot}
\alias{nlme.vpcplot}
\title{Plot a VPC of an nlme object}
\usage{
\method{nlme}{vpcplot}(
  object,
  vpc,
  formula,
  obsData = NULL,
  xLabel,
  yLabel,
  xLimits,
  yLimits,
  logY = FALSE,
  color.style = list(observed = rgb(0.2, 0.2, 0.2), central = blue[10], inner =
    blue[4], outer = blue[1]),
  symbol = list(pch = 15, cex = 0.5),
  aspect = 1,
  cex.label = 1.5,
  nx = NULL,
  showPredAs = "area",
  showObsDots = TRUE,
  showObsLines = TRUE,
  obscol.dot = gray[8],
  obscex.dot = 0.5,
  obspch.dot = 1,
  obscol.line = gray[10],
  predcol.central = blue[6],
  predcol.outer = blue[7],
  predcol.area = blue[1],
  predcol.inner = blue[3],
  ...
)
}
\arguments{
\item{object}{the nlme object (if it's an nlme.run object the input should read \code{object$object})}

\item{vpc}{the output from \code{nlme.vpc}}

\item{formula}{a formula detailing what to plot}

\item{obsData}{the observation data, if not supplied defaults to \code{obsData = getData(object)}. Use if newdata in extract.nlme was specified}

\item{xLabel}{same as xlab}

\item{yLabel}{same as ylab}

\item{xLimits}{same as xlim}

\item{yLimits}{same as ylim}

\item{logY}{logical to log Y axis}

\item{color.style}{list with layout options, such as:}

\item{symbol}{list with elements \code{pch} (15) and \code{cex} (0.5)}

\item{aspect}{trellis banking value (defaults to 1 for sqaure panels)}

\item{cex.label}{tex size for x and y labels (1.5)}

\item{nx}{number of bins for x covariate}

\item{showPredAs}{layout style (defaults to "area" for polygons but could be 'lines' too).}

\item{showObsDots}{logical to showobserved dots (T)}

\item{showObsLines}{logical to show quantile lines of observed (T)}

\item{obscol.dot}{color of observed dot}

\item{obscex.dot}{symbol size}

\item{obspch.dot}{type of symbol}

\item{obscol.line}{observed line color}

\item{predcol.central}{predicted central tendency color}

\item{predcol.outer}{predicted outer polygon color}

\item{predcol.area}{predicted overall polygon color}

\item{predcol.inner}{predicted inner polygon (IQR) color}

\item{...}{any additional arguments passed on to the lattice call}
}
\value{
A VPC plot.
}
\description{
Plot a VPC of an nlme object
}
\examples{
# library(nlme)
pkpdData = example.pkpdData()
EFF.1comp.1abs <- function(dose, tob, cl, v, ka, keo)
 {
   # Effect-site concentration for 1-compartment model, 1st-order absorption

   kel = cl / v

   # Define coefficients
   A = 1/(kel-ka) / (keo-ka)
   B = 1/(ka-kel) / (keo-kel)
   C = 1/(ka-keo) / (kel-keo)

   # Return effect-site concentration
   dose*ka*keo/v * (A*exp(-ka*tob) + B*exp(-kel*tob) + C*exp(-keo*tob))
 }
 fit.PD004.nlme = nlme.run(
   model = value ~ base + EFF.1comp.1abs(dose, time, cl * exp(cl.eta), v, ka, keo),
   data = pkpdData[pkpdData$type == "PD" & pkpdData$dose > 0 & pkpdData$value > 0.5, ],
   fixed = base + cl + v + ka + keo ~ 1,
   random = cl.eta ~ 1,
   groups = ~ id,
   start = c(base = 1, cl = 1, v = 10, ka = 1, keo = 0.01),
   problem = "True Model",
   reference = 4)
 summary(fit.PD004.nlme$object)
 nlme.extract(fit.PD004.nlme$object)$table
 vpc.PD004.nlme = nlme.vpc(fit.PD004.nlme$object, nrep = 100)
 nlme.vpcplot(fit.PD004.nlme$object, vpc.PD004.nlme)
}
\seealso{
\code{\link{nlme.vpc}}, \code{\link{nlme.extract}}
}
