% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nls.predict.r
\name{nls.predict}
\alias{nls.predict}
\title{Predict partial residuals from an nls object}
\usage{
nls.predict(func, object, newdata = NULL, xrange = 150)
}
\arguments{
\item{func}{formula of \code{y ~ x | z} detailing what to predict. Requires \code{y} to be the response variable of the nls obhect}

\item{object}{and nls object}

\item{newdata}{optional dataset to simulate across. Needs the elements required to create the prediction}

\item{xrange}{number of equal-spaced x covariate values created from \code{min(x)} to \code{max(x)}}
}
\value{
a graph or a multi-level list with observed and predicted output
}
\description{
Predict partial residuals from an nls object to be plotted with \code{plot.fit}
}
\examples{
DNase1 <- subset(DNase, Run == 1)

## using a selfStart model
fm1DNase1 <- nls(density ~ SSlogis(log(conc), Asym, xmid, scal), DNase1)
summary(fm1DNase1)

fm1DNase1.predict = nls.predict(density ~ conc,object = fm1DNase1)
plot.fit(fm1DNase1.predict)
fm1DNase1.predict = nls.predict(density ~ conc,object = fm1DNase1)
plot.fit(fm1DNase1.predict, logX = T)
}
\seealso{
\code{\link{plot.fit}}, \code{\link{nlme.predict}}
}

