% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nm.compare.plot.r
\name{nm.compare.plot}
\alias{nm.compare.plot}
\title{Create a runrecord like GOF plot of selected runs}
\usage{
nm.compare.plot(runs, path = getOption("nmDir"), keep.cols = Cs(ID, DV,
  CWRES, PRED, IPRED, TIME, EVID), alias = list(DV = "CONC", TIME = "TAFD"),
  dot.size = 1, text.size = 1, log = F)
}
\arguments{
\item{runs}{character vector with rootname runs like \code{c("run1","run2")}}

\item{path}{directory where \code{runs} reside}

\item{keep.cols}{needs at the minimum \code{Cs(ID,DV,CWRES,PRED,IPRED,TIME,EVID)} but can be altere to match other data columns}

\item{alias}{a named list to rename DV and TIME variables, in case in \code{$DATA} they have been specified as \code{...=DV} and/or \code{...=TIME}}

\item{dot.size}{scalar for GOF plot dots}

\item{text.size}{scalar for text size in strip labels}

\item{log}{a logical indicating if the GOF plots need to be logged}
}
\value{
A plot
}
\description{
Create a runrecord like GOF plot of selected runs
}
\examples{
nm.compare.plot(runs = c("example1","example2")
, path =getOption("qpExampleDir")
, alias = list(DV = "CONC",TIME = "TIME")
, keep.cols = Cs(ID,DV,CWRES,PRED,IPRED,TIME,EVID)
, text.size = 0.6
)
}

