% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nm.cormat.extract.r
\name{nm.cormat.extract}
\alias{nm.cormat.extract}
\title{Correlation Matrix Extraction}
\usage{
nm.cormat.extract(run, path = getOption("nmDir"))
}
\arguments{
\item{run}{character string in the form of 'run1'}

\item{path}{Directory where .cov file resides}
}
\value{
Covariate matrix
}
\description{
Extract correlation matrix from a .cor file
}
\examples{
nm.unzip(run="example1", extension=".cor", path = file.path(getOption("qpExampleDir"),"example1"))
run1.cormat =  nm.cormat.extract("example1", path = getOption("qpExampleDir"))
names(run1.cormat)
names(run1.cormat[[1]])
lapply(run1.cormat, dim)
file.remove( file.path(getOption("qpExampleDir"),"example1/example1.cor"))
}
\seealso{
\code{\link{extract.varcov}}, \code{\link{get.covmat}}, \code{\link{nm.covmat.extract}}
}

