% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.covmat.extract.r
\name{nm.covmat.extract}
\alias{nm.covmat.extract}
\title{Covariance Matrix Extraction}
\usage{
nm.covmat.extract(run, path = getOption("nmDir"))
}
\arguments{
\item{run}{character string in the form of 'run1'}

\item{path}{Directory where .cov file resides}
}
\value{
Covariate matrix
}
\description{
Extract covariance matrix from a .cov file
}
\examples{
nm.unzip(run = "example1", extension = ".cov", path = file.path(getOption("qpExampleDir"),"example1"))
run1.covmat =  nm.covmat.extract("example1", path = getOption("qpExampleDir"))
names(run1.covmat)
names(run1.covmat[[1]])
lapply(run1.covmat, dim)
file.remove( file.path(getOption("qpExampleDir"),"example1/example1.cov"))

}
\seealso{
\code{\link{extract.varcov}}, \code{\link{get.covmat}}, \code{\link{nm.cormat.extract}}
}
