% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.create.control.r
\name{nm.create.control}
\alias{nm.create.control}
\title{Documentation stub
Create a control stream for simulation}
\usage{
nm.create.control(run, path = getOption("nmDir"),
  outputFile.extension = ".out", modelFile.extension = ".ctl",
  seed = 1234567, newDataFile = NULL, uncertainty = FALSE,
  sim = NULL, table = NULL)
}
\arguments{
\item{run}{run rootname (e.g. \code{run1})}

\item{path}{directory where \code{run} resides}

\item{outputFile.extension}{file extension of the NONMEM output file. Defaults to ".lst"}

\item{modelFile.extension}{file extension of the NONMEM control stream. Defaults to ".mod"}

\item{seed}{seed number to go into \code{$SIM}}

\item{newDataFile}{if some else as the model data file should be used}

\item{uncertainty}{logical for simulation across uncertainty}

\item{sim}{to be described}

\item{table}{a character string with the required \code{$TABLE} statement for the simulation control stream}
}
\value{
A control stream ready for execution written to file
#' @note THIS FUNCTION HAS NOT BEEN THOROUGHLY TESTED. USE WITH CAUTION AND ADAPT AS NEEDED ON THE PROJECT LEVEL.
}
\description{
Create a control stream for simulation with sampled TH/OM/SG estimates
}
