% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.extract.ext.r
\name{nm.extract.ext}
\alias{nm.extract.ext}
\title{NONMEM output extraction using .lst and .ext files}
\usage{
nm.extract.ext(run, path = getOption("nmDir"))
}
\arguments{
\item{run}{run rootname (e.g. \code{run1})}

\item{path}{directory where \code{run} resides}
}
\value{
A list with the NONMEM version, OFV, the xpose tables and the parameter table
}
\description{
Extract NONMEM output based on the .ext output file
}
\examples{
run2.ext = nm.extract.ext("example2", path = getOption("qpExampleDir"))
run2.ext$version; run2.ext$final_objective_function
}
\seealso{
\code{\link{nm.extract.xml}}, \code{\link{nm.params.table}}, \code{\link{get.nm.version}}, \code{\link{get.ofv}}, \code{\link{get.xpose.tables}}
}
