% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.extract.xml.r
\name{nm.extract.xml}
\alias{nm.extract.xml}
\title{NONMEM output extraction from XML output}
\usage{
nm.extract.xml(
  run = "run1",
  path = getOption("nmDir"),
  filename = paste(path, run, paste(run, xml.extension, sep = ""), sep = "/"),
  digits = 3,
  digits.se = digits,
  digits.cv = digits,
  digits.est = digits,
  xml.extension = ".xml",
  zip.extension = ".7z",
  zip.call = getOption("unzip.call"),
  remove.obsolete = FALSE,
  na.value = "n.d.",
  control_stream = FALSE,
  get.xpose.tables = TRUE,
  quiet = TRUE
)
}
\arguments{
\item{run}{run rootname (e.g. \code{run1})}

\item{path}{directory where \code{run} resides}

\item{filename}{with full path of the XML file to be parsed. This is automatically created using qP workflow standards but can be specified as well}

\item{digits}{number of significant digits}

\item{digits.se}{number of significant digits in SE (when point estimates = T)}

\item{digits.cv}{number of significant digits in CV\% (when point.estimates = T)}

\item{digits.est}{number of significant digits in parameter estimates (when point.estimates = T)}

\item{xml.extension}{extension of the xml file. Defaults to ".xml"}

\item{zip.extension}{extension of the zip program. Defaults to ".7z"}

\item{zip.call}{full OS call to zip utility for unzipping. Defaults to "c:/progra~1/7-zip/7z e \%s.7z" on windows}

\item{remove.obsolete}{logical (F) defining if non-estimated parameters should be dropped for the parameter table.}

\item{na.value}{what to do with an NA value in the parameter table. Defaults to "n.d."}

\item{control_stream}{logical (F) defining if the control stream should be included as well.}

\item{get.xpose.tables}{logical (T) defining if the xpose tables should be integrated in the result as a single data frame using the function \code{get.xpose.tables}}

\item{quiet}{if TRUE (default) silences any message returned by the function and subfunctions.}
}
\value{
An extensive list with all elements extracted from the XML output
}
\description{
Extracts all output embedded in the XML file of a NONMEM run
}
\note{
Works for NM7.2 (up to 9 OMEGAs) and NM7.3 (bug free)
}
\examples{
tmp = nm.extract.xml("example2", path = getOption("qpExampleDir"))
names(tmp)
tmp$final_objective_function
tmp$table
tmp$monitor
tmp = nm.extract.xml("example2", path = getOption("qpExampleDir"))
}
