% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.insert.dose.multiple.compartments.r
\name{nm.insert.dose.multiple.compartments}
\alias{nm.insert.dose.multiple.compartments}
\title{Insert doses in other compartments}
\usage{
nm.insert.dose.multiple.compartments(data, dose.in.cmt = c(1, 2), ...,
  quiet = TRUE)
}
\arguments{
\item{data}{NONMEM data.frame, should contain variable 'CMT'}

\item{dose.in.cmt}{numeric double or vector for new CMT doses}

\item{...}{(unquoted) columns to sort the result by}

\item{quiet}{when TRUE (default) silences the messages returned}
}
\value{
A data.frame with input dataset supplied with additional doses
}
\description{
To model drug entry into the system from multiple absorption sites this function allows to turn a NONMEM datset into one with additional doses in differetn compartments.
}
\examples{
library(dplyr)
library(Hmisc)
## original data:
nmData = example.NONMEM.dataset(TIME=seq(0,24,4))
nmData$CMT = with(nmData, swap(EVID, 0:1, 2:1))

## adding doses to compartments 3, 4 and 5
nmData3 = nm.insert.dose.multiple.compartments(data = nmData
                                               , dose.in.cmt=c(3,4,5)
                                               , ID, TIME, -EVID
                                               
)
tbl_df(nmData3[, Cs(ID,TIME,EVID,AMT,CMT,DV)])
}
\seealso{
\code{\link{nm.insert.evid2amt0}}
}
