% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.modeltrail.graph.r
\name{nm.modeltrail.graph}
\alias{nm.modeltrail.graph}
\title{Graphical Bayesian model trail}
\usage{
nm.modeltrail.graph(runs, path = getOption("nmDir"), file.ext = ".ext",
  col.area = gray[5], col.mean = red[7], col.percentiles = gray[5],
  relabel = NULL, strip.size = NULL, ...)
}
\arguments{
\item{runs}{charater vector with run rootname (e.g. \code{c("run1","run2","run3")})}

\item{path}{directory where \code{runs} reside}

\item{col.area}{polygon color of the density plot}

\item{col.mean}{color of the mean line}

\item{col.percentiles}{color of the percentile lines}

\item{relabel}{user-specified labels appearing in the Lattice strips}

\item{strip.size}{scalar to for the strip size in case strip text is too small ortoo large}

\item{...}{graphical arguments passed on to the lattice plots}

\item{extension}{file extension of the run.ext file. Defaults to ".ext"}
}
\value{
Lattice graph with the distributions of the MCMC Bayesian OFV, supplied with mean line and 2.5th and 97.5th percentiles
}
\description{
Create graphical presentation of Bayesian model building trail. The function reads the iterations from the run.ext file and displays them as densities in a lattice graph
}
\examples{
## when there is no Bayesian estimation
nm.modeltrail.graph(runs = c("example1","example2")
                    , relabel = c("structural model", "final covariate model")
                    , file.ext = ".mext"
                    , path = getOption("qpExampleDir")
)
}
\seealso{
\code{\link{read.ext}}
}
