% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.params.table.r
\name{nm.params.table}
\alias{nm.params.table}
\title{Create NONMEM Parameter Estimate Table}
\usage{
nm.params.table(run, path = getOption("nmDir"), file.ext = ".ext",
  runIndex, fixed.text = ".....", return.all = FALSE)
}
\arguments{
\item{run}{run rootname (e.g. run1)}

\item{path}{directory where rootname.ext resides}

\item{file.ext}{file extentions of the .ext file. Defaults to .ext}

\item{runIndex}{which (numeric) estimation method needs to be tabulated? Defaults to the last estimation method outputted to the .ext file}

\item{fixed.text}{charactre string to replace the entries for standard error and coefficient of variation (CV) for parameters that were fixed in the control stream}

\item{return.all}{logical indicator to return the standard data set for inclusion in reports (default) or if TRUE the data set with everything.}
}
\value{
data frame with parameter name, estimate, coefficient of variation, standard error, and estimated/fixed information.
}
\description{
Create NONMEM Parameter Estimate Table
}
\examples{
nm.params.table(run = "example1", path = getOption("qpExampleDir"))
nm.params.table("example2",  path = getOption("qpExampleDir"), fixed.text = "(fixed to 0)", return.all = TRUE)
}
\seealso{
\code{\link{process.parTable}}
}
