% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.parse.control.stream.r
\name{nm.parse.control.stream}
\alias{nm.parse.control.stream}
\title{Parse NONMEM control stream}
\usage{
nm.parse.control.stream(run, path = getOption("nmDir"),
  file.ext = ".mod", quiet = TRUE)
}
\arguments{
\item{run}{run rootname (e.g. \code{run1})}

\item{path}{directory where \code{run} resides}

\item{quiet}{if TRUE no message will be returned.}

\item{extension}{file extension of NONMEM control stream (.mod)}
}
\value{
A named list with all \code{$PK, $ERROR, $THETA} etc... elements of the control stream
}
\description{
Function to read NONMEM control stream and parse it out as named list
}
\examples{
nm.parse.control.stream("example1", path = getOption("qpExampleDir"), file.ext = ".ctl")

}
