% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nm.parse2r.operators.r
\name{nm.parse2r.operators}
\alias{nm.parse2r.operators}
\title{Change NONMEM operators into R operators}
\usage{
nm.parse2r.operators(nmtext)
}
\arguments{
\item{nmtext}{NONMEM text like 'DOSE.GT.10'.}
}
\value{
Character string with an R expressions of the NONMEM input
}
\description{
Change NONMEM operators into R operators
}
\note{
This will only with for simple operators but does not yet accomodate (nested) IF-THEN-ELSE blocks.
}
\examples{
nm.parse2r.operators(c("DOSE.GT.10","AMT.LE.10","ID.EQ.1000","SEX!=0"))
}

