% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.process.coveffects.r
\name{make.eqs}
\alias{make.eqs}
\alias{eval.eqs}
\alias{nm.process.coveffects}
\title{Make a list of equations}
\usage{
make.eqs(...)

eval.eqs(.data, ...)

nm.process.coveffects(eqs, covs.info, pars, xpose.df, omega,
  Nboot = 10000, pi.wid = 0.9)
}
\arguments{
\item{...}{A series of named formulas}

\item{.data}{is data.frame, list, or environment}

\item{eqs}{A list of calls}

\item{covs.info}{A list of covInfo objects to parameterize the subanalysis}

\item{xpose.df}{A dataframe containing post-hoc parameter values}

\item{omega}{The random effect covariance matrix from which to simulate ETA values}

\item{Nboot}{The number of bootstrap samples to run for each sample point}

\item{pi.wid}{The width of prediction intervals (e.g. 0.9 is a 90\% prediction interval)}

\item{...}{is a list of calls or expressions}
}
\value{
A list of call objects that can be used to evaluate the input equations 
in a given environment.

.data is returned with the evaluated expressions added

A list including lists of: covariate effect plots, table of covariate effects,
table of post-hoc values -- for each covariate effect; the call list; the list of covInfo
}
\description{
A list of equations that define the covariate relationships is...
}
\section{Functions}{
\itemize{
\item \code{make.eqs}: Makes a list of calls from its named arguments 
input formulas.

\item \code{eval.eqs}: Evaluate a list of calls in an environment, list, 
or data.frame.
}}

\note{
This function is used with \code{nm.process.coveffects}

This function is used with \code{nm.process.coveffects}
}
\examples{
eqs=make.eqs(CL=THETA1*(1 + (GNDR==0)*THETA7)*(AGE/33.72)^THETA6*exp(ETA1)
            ,V1=THETA2*(1 + (GNDR==0)*THETA9)*(AGE/33.72)^THETA8*exp(ETA2)
            )
eqs$label=list(CL="Clearance (L/h)", V1="Volume of Central Compartment (L)")            

eqs=make.eqs(CL=THETA1*(1 + (GNDR==0)*THETA7)*(AGE/33.72)^THETA6*exp(ETA1)
            ,V1=THETA2*(1 + (GNDR==0)*THETA9)*(AGE/33.72)^THETA8*exp(ETA2)
            )
eqs$label=list(CL="Clearance (L/h)", V1="Volume of Central Compartment (L)")            

}
