% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nm.remove.section.r
\name{nm.remove.section}
\alias{nm.remove.section}
\title{Extract control stream sections}
\usage{
nm.remove.section(secNames, ctl)
}
\arguments{
\item{secNames}{character vector with sections to remove}

\item{ctl}{character vector of scanned control stream}
}
\value{
A list with as many (named) element as there are $.... elements in the control stream
}
\description{
Function to read NONMEM control stream and extract specific sections
}
\note{
This function is not intended to be called alone, but is called in the wrapper function \code{nm.parse.control.stream}
}
\examples{
nm.remove.section("EST"
   , read.mod("example1", path = getOption("qpExampleDir"), file.ext=".ctl")
   )[[2]]
}
\seealso{
\code{\link{nm.parse.control.stream}}
}

