% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.runtime.r
\name{nm.runtime}
\alias{nm.runtime}
\title{NONMEM run time}
\usage{
nm.runtime(run, path = getOption("nmDir"))
}
\arguments{
\item{run}{run rootname (e.g. \code{'run1'})}

\item{path}{directory where \code{run} resides}
}
\description{
Get the NONMEM run time
}
\examples{
nm.runtime(run = "example1", path = getOption("qpExampleDir"))
}
\seealso{
\code{\link{get.nm.version}}, \code{\link{get.shrinkage}}, \code{\link{get.ofv}}, \code{\link{read.out}}
}
