% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.vpcplot.r
\name{nm.vpcplot}
\alias{nm.vpcplot}
\title{Create covariate plot and data}
\usage{
nm.vpcplot(
  path,
  result = "vpc_results.csv",
  tab = dir(path = path, pattern = "^vpctab")[1],
  PI.limits = c(0.025, 0.975),
  subset = NULL,
  output = FALSE,
  relabel.strata = NULL,
  subtitle = "",
  logY = FALSE,
  logX = FALSE,
  xLabel = NULL,
  yLabel = NULL,
  cex.label = 1.25,
  cex.strip.labels = 1,
  xLimits = NULL,
  yLimits = NULL,
  aspect = 1,
  scales = NULL,
  xscale.components = NULL,
  yscale.components = NULL,
  layout,
  showPredAs = "area",
  showObsDots = TRUE,
  showObsLines = TRUE,
  lineType = "l",
  central.lwd = 1.5,
  obscol.dot = gray[8],
  obscex.dot = 0.5,
  obspch.dot = 1,
  obscol.line = gray[10],
  predcol.central = red[6],
  predcol.outer = blue[7],
  predcol.area.central = red[2],
  predcol.area.outer = blue[1],
  predcol.area = blue[1]
)
}
\arguments{
\item{path}{directory where vpc_results and vpctab files reside}

\item{result}{filename of the 'vpc_results.csv" file}

\item{tab}{filename of the 'vpctab' file}

\item{PI.limits}{empty}

\item{subset}{empty}

\item{output}{empty}

\item{relabel.strata}{empty}

\item{subtitle}{empty}

\item{logY}{empty}

\item{logX}{empty}

\item{xLabel}{empty}

\item{yLabel}{empty}

\item{cex.label}{empty}

\item{cex.strip.labels}{empty}

\item{xLimits}{empty}

\item{yLimits}{empty}

\item{aspect}{empty}

\item{scales}{empty}

\item{xscale.components}{empty}

\item{yscale.components}{empty}

\item{layout}{empty}

\item{showPredAs}{empty}

\item{showObsDots}{empty}

\item{showObsLines}{empty}

\item{lineType}{= "l",empty}

\item{central.lwd}{empty}

\item{obscol.dot}{empty}

\item{obscex.dot}{empty}

\item{obspch.dot}{empty}

\item{obscol.line}{empty}

\item{predcol.central}{empty}

\item{predcol.outer}{empty}

\item{predcol.area.central}{empty}

\item{predcol.area.outer}{empty}

\item{predcol.area}{empty}
}
\value{
a lattice VPC graph
}
\description{
Create covariate plot and data
}
\note{
This function should not be used as it was marked for future deprecation. qPharmetra's default VPC plot tool is xpose.VPC
}
