% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_unzip.r
\name{nm_unzip}
\alias{nm_unzip}
\title{Unzip a NONMEM File}
\usage{
nm_unzip(
  run,
  extension = ".xml",
  filename = paste0(run, fxs(extension, ".   ")),
  path = file.path(normalizePath(getOption("nmDir", getwd())), run),
  tmpdir = tempdir(),
  temp = FALSE,
  outdir = ifelse(temp, tmpdir, path),
  zip.filename = file.path(path, filename),
  zip.call = getOption("unzip.call"),
  quiet = TRUE,
  ...
)
}
\arguments{
\item{run}{character (run name)}

\item{extension}{character: extension to append to run}

\item{filename}{character: supersedes run, extension}

\item{path}{path to directory enclosing the NONMEM run directory}

\item{outdir}{character: one of \code{path} or \code{tempdir}}

\item{zip.call}{character: system unzip invocation}

\item{tempdir}{character: path to temporary working directory if \code{temp} is true}

\item{temp:}{whether to unzip to a temporary directory}

\item{zip.filename:}{character: path and source filename, supersedes extension, filename, and path}
}
\description{
Unzips a NONMEM file, by default the xml output
of a NONMEM run.  Uses 7z decompression.
}
\concept{nm}
\keyword{internal}
