% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odeSim.R
\name{odeSim}
\alias{odeSim}
\title{ODE simulation with Multiprocessor Computation}
\usage{
odeSim(state, times, func, parms, simPar, simParValues, method = "lsoda",
  nodes = 1, keepTimeDT = 0.1)
}
\arguments{
\item{state}{initial conditions}

\item{times}{at what times should the system of differential equations be updated}

\item{func}{function describing the system of differential equations}

\item{parms}{global parameter values}

\item{simPar}{names of parameters and states that should be changed from global values}

\item{simParValues}{values of parameters and states - each row corresponds to one scenario}

\item{method}{method for updating the states of the differential equations (see ?ode)}

\item{nodes}{number of parallel nodes to use in computations}

\item{keepTimeDT}{The times that return a zero when modulus (divided) by keepTimeDT are kept}
}
\value{
Data.frame with simulated output
}
\description{
Simulate with ordinary differential equations in R including multiprocessor computation
}
