% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.avg.LP.r
\name{panel.avg.LP}
\alias{panel.avg.LP}
\title{Lattice panel function for average lines}
\usage{
panel.avg.LP(
  x,
  y,
  fun = mean,
  horizontal = TRUE,
  lwd = reference.line$lwd,
  lty = reference.line$lty,
  col,
  col.line = reference.line$col,
  type = "l",
  ...,
  identifier = "linejoin",
  show.points = FALSE
)
}
\arguments{
\item{x, y}{The regular x, y input for panel functions}

\item{fun}{function to summarize by, defaults to \code{mean}}

\item{horizontal}{logical (T) indicating if the average needs to be taken vertical or horizontal}

\item{lwd}{line width (taken from \code{trellis.par.get("reference.line")})}

\item{lty}{line type (taken from \code{trellis.par.get("reference.line")})}

\item{col}{observed data color}

\item{col.line}{line color (taken from \code{trellis.par.get("reference.line")})}

\item{type}{same as type in a lattice call (default is "l" for lines)}

\item{\dots}{passed to panel.points() and panel.lines()}

\item{identifier}{option to join lines (which is default)}

\item{show.points}{logical indicating to show observed data dots}
}
\value{
Nothing -> internal to a lattice call
}
\description{
Panel function that can be used with xyplot() (lattice) and related functions to plot the average (or other metric) of y values at each X, while also drawing connecting lines and points. This is modified version of panel.average, and is used by \code{panel.meanspag()}
}
