% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/panel.cwres.r
\name{panel.cwres}
\alias{panel.cwres}
\title{Lattice panel function for CWRES plots}
\usage{
panel.cwres(x, y, ..., col = obs.color, col.line = abline.color, pch = 1,
  col.loess = smooth.color)
}
\arguments{
\item{x,y}{numeric vectors x and y}

\item{...}{any other arguments to be passed on to the lattice call}

\item{col}{symbol color (defaults to obs.color)}

\item{cex}{symbol size}

\item{smooth.col}{color of smooth line}

\item{lwd.loess}{line width of loess smooth}
}
\value{
Lattice panel output (invisible)
}
\description{
Panel function to plot CWRES vs any covariate
}
\examples{
out = get.xpose.tables("example2", getOption("qpExampleDir"))
 trellis.strip.color()
xyplot(CWRES ~ value | variable
   , data = subset(reshape2::melt(out, measure.vars = Cs(PRED,TIME)),EVID==0)
   , aspect = 1
   , scales = list(x = list(relation = "free"))
   , panel = panel.cwres
   )
}

