% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.modelfit.r
\name{panel.modelfit}
\alias{panel.modelfit}
\title{Panel function for Residual Variability Plots}
\usage{
panel.modelfit(
  x,
  y,
  groups,
  subscripts,
  logX = FALSE,
  logY = FALSE,
  type.obs = "b",
  pred.lty = "solid",
  ipred.lty = "solid",
  ...
)
}
\arguments{
\item{x, y}{numeric vectors x and y}

\item{groups}{see \code{\link[lattice]{xyplot}}}

\item{subscripts}{see \code{\link[lattice]{xyplot}}}

\item{logX}{log the x-axis (defaults to F)}

\item{logY}{log the y-axis (defaults to F)}

\item{type.obs}{line type for observed data (defaults to 'b' for both lines and points)}

\item{pred.lty}{line type for population predictions}

\item{ipred.lty}{line type for individual predictions}

\item{\dots}{passed to llines()}
}
\value{
Lattice panel output (invisible)
}
\description{
Panel function to plot DV, PRED, IPRED overlays for individual profiles
}
\examples{
library(Hmisc)
out = get.xpose.tables('example1',getOption('qpExampleDir'))
out$EVID=0
trellis.strip.color()
xyplot(Cbind(CONC,PRED,IPRED,EVID) ~ TIME
       , groups = ID
       , data = subset(out)
       , scales = list(x = list(relation = 'free'),y = list(log=10))
       , panel = panel.modelfit
       , logY = TRUE
       , yscale.components = yscale.components.log10
)
}
