% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/panel.nlme.vpc.obs.r
\name{panel.nlme.vpc.obs}
\alias{panel.nlme.vpc.obs}
\title{Lattice function for observed components of an nlme VPCs}
\usage{
panel.nlme.vpc.obs(x, y, logY, showPredAs, showObsDots, showObsLines,
  col.scheme, obscex.dot, obspch.dot, ...)
}
\arguments{
\item{other}{a collection of parameters are available to modify the layout of the VPC}
}
\value{
Nothing -> internal to a lattice call
}
\description{
Panel function for observed lines / colored polygons of a vpc plot of an nlme object
}
\note{
Not to be used as standalone. Will be be called by \code{nlme.vpcplot}
}
\seealso{
\code{\link{nlme.vpcplot}}
}

