% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.nlme.vpc.pred.r
\name{panel.nlme.vpc.pred}
\alias{panel.nlme.vpc.pred}
\title{Lattice function for predicted components of an nlme VPCs}
\usage{
panel.nlme.vpc.pred(x, y, OBS, vpc, subscripts, logY, showPredAs,
  showObsDots, showObsLines, xCov, col.scheme, obscex.dot, obspch.dot, ...)
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}

\item{OBS}{unused}

\item{vpc}{unused}

\item{subscripts}{unused}

\item{logY}{whether to log-transform the y axis}

\item{showPredAs}{choice of 'lines' or 'area'}

\item{showObsDots}{unused}

\item{showObsLines}{unused}

\item{xCov}{unused}

\item{col.scheme}{list with elements: [pred: [outer, area, inner]]}

\item{obscex.dot}{unused}

\item{obspch.dot}{unused}

\item{\dots}{passed to llines()}
}
\value{
Lattice panel output (invisible)
}
\description{
Panel function for predicted lines / colored polygons of a vpc plot of an nlme object
}
\note{
Not to be used as standalone. Will be be called by \code{nlme.vpcplot}
}
\seealso{
\code{\link{nlme.vpcplot}}
}
