% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.nonmem.vpc.r
\name{panel.nonmem.vpc}
\alias{panel.nonmem.vpc}
\title{Lattice function for predicted components of a NONMEM VPC}
\usage{
panel.nonmem.vpc(x, y, OBS, vpc, subscripts, logY, logX, showPredAs,
  showObsDots, showObsLines, xCov, col.scheme, obscex.dot, obspch.dot,
  lineType, central.lwd, ...)
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}

\item{OBS}{list with element: strata_no}

\item{vpc}{list with element: strata_no}

\item{subscripts}{subscripts for vpc}

\item{logY}{logical for logging Y coordinate}

\item{logX}{logical for logging X coordinate}

\item{showPredAs}{choice of 'lines', 'area', or 'xpose'}

\item{showObsDots}{whether to show observations as dots}

\item{showObsLines}{whether to show observations as lines}

\item{xCov}{covariates of interest from vpc}

\item{col.scheme}{list with elements: [pred: [outer, central, area], obs: [dot, line]]}

\item{obscex.dot}{cex for observations}

\item{obspch.dot}{pch for observations}

\item{lineType}{line type for observations and predictions}

\item{central.lwd}{line width for predicted central tendency}

\item{\dots}{passed to panel.xyplot() and llines()}
}
\value{
Lattice panel output (invisible)
}
\description{
Panel function for predicted and observed lines / colored polygons of a vpc plot of an NONMEM / PsN VPC object
}
\note{
Not to be used as standalone. Will be be called by \code{nm.vpcplot}. Furthermore, the \code{nm.vpcplot} routine will go deprecated soon, thus it is not adviced to use this function going forward. Instead use xpose.VPC.
}
\seealso{
\code{\link{nm.vpcplot}}
}
