% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.residual.r
\name{panel.residual}
\alias{panel.residual}
\title{Panel function for Residual Variability Plots}
\usage{
panel.residual(
  x,
  y,
  abline = "identity",
  col = obs.color,
  cex = 1,
  span = 0.8,
  lwd.loess = 2,
  smooth.col = smooth.color,
  ...
)
}
\arguments{
\item{x, y}{numeric vectors x and y}

\item{abline}{if "identity" (default) the line of identity is drawn. If anything else no abline is drawn.}

\item{col}{symbol color (defaults to obs.color)}

\item{cex}{symbol size}

\item{span}{passed to panel.loess()}

\item{lwd.loess}{line width of loess smooth}

\item{smooth.col}{color of smooth line}

\item{\dots}{passed to panel.xyplot(), panel.abline(), and panel.loess()}
}
\value{
Lattice panel output (invisible)
}
\description{
Panel function to plot residuals vs any covariate or DV vs (I)PRED
}
\examples{
library(lattice)
library(reshape2)
trellis.strip.color()
xyplot(CONC ~ value | variable
       , data = subset(
          reshape2::melt(
             get.xpose.tables('example1', getOption('qpExampleDir')
             )
             , measure.vars = c('PRED','IPRED')
          ), EVID == 0)
       , subset = EVID == 0
       , aspect = 1
       , scales = list(x = list(relation = 'free'))
       , panel = panel.residual
       )


}
