% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.superpose.bubble.r
\name{panel.superpose.bubble}
\alias{panel.superpose.bubble}
\title{Panel function for bubble plots}
\usage{
panel.superpose.bubble(
  x,
  y,
  n,
  groups,
  subscripts,
  cols = NULL,
  syms = NULL,
  ...
)
}
\arguments{
\item{x, y}{numeric vectors x and y}

\item{n}{the column containing the sample ie of each record}

\item{groups}{grouping levels}

\item{subscripts}{subscripts}

\item{cols}{bubble colors}

\item{syms}{bubbles symbol types}

\item{\dots}{passed to llines() and lpoints()}
}
\value{
Lattice panel output (invisible)
}
\description{
Panel function to create bubble plots where the symbol size varyies in proportion to a variable
}
\examples{
library(lattice)
myData = data.frame(  n = c(5,100,332,12,50,74)
, value = c(0.13,0.78,0.49,0.67,0.34,0.57)
, dose = c(20,25,50,30,40,35)
, sex = rep(c("M","F"), ea = 3)
)
myData = expand.data(myData, values = c(0,1,2,3,4), name = "time")
myData$ypred = myData$value * myData$time ^3/(1 + myData$time^3)
xyplot(ypred ~ time
       , data = myData
       , groups = dose
       , subscripts = TRUE
       , nn = sqrt(myData$n/10)
       , panel = function(x,y, nn = nn, groups, subscripts,...)
       {
         panel.superpose.bubble(x,y,nn, groups, subscripts, ...)
       }
)
}
