% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.thalf.r
\name{panel.thalf}
\alias{panel.thalf}
\title{Panel function for elimination half life}
\usage{
panel.thalf(
  x,
  y,
  ...,
  text.cex = 0.8,
  line.col = "blue",
  nSignif = 3,
  unit = "hr"
)
}
\arguments{
\item{x, y}{numeric vectors x and y}

\item{...}{any other arguments passed on to the lattice call}

\item{text.cex}{text font size}

\item{line.col}{line color}

\item{nSignif}{number of significant digits in output}

\item{unit}{time unit (hr)}
}
\value{
Lattice panel output (invisible)
}
\description{
Panel function to create individual plots with terminal half life predictions
}
\examples{
library(lattice)
pkpdData = example.pkpdData()
pkpdData$blq = 0
pkpdData$blq[pkpdData$type=="PK"&pkpdData$value<0.1] = 1
xyplot(value ~ time | id
, data = subset(pkpdData, blq==0)
, subset = type == "PK" & dose == 100 & value > 1e-6
, groups = id
, panel = function(x,y, ...)
{
  panel.xyplot(x,y,..., col = "gray", pch = 18, type = "b", cex = 1)
  sel = length(x) 
  sel = seq(sel-3,sel)
  panel.thalf(x[sel],y[sel],...,lwd = 2)
  lpoints(x,y, col = steel)
}
, scales = list(y = list(log = 10))
, yscale.components = yscale.components.log10
)
}
