% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qP.colors.r
\docType{data}
\name{PI.ci.up.arcol}
\alias{PI.ci.up.arcol}
\title{PI.ci.up.arcol (Autograph)}
\format{
An object of class \code{character} of length 1.
}
\usage{
PI.ci.up.arcol
}
\description{
PI.ci.up.arcol (Autograph)
}
\seealso{
Other qpcolors (Autograph): 
\code{\link{loi.color}}
}
\concept{qpcolors (Autograph)}
\keyword{datasets}
