% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pk.pred.r
\name{pk.pred}
\alias{eff.1comp.1abs}
\alias{eff.1comp.1abs.ss}
\alias{eff.1comp.iv}
\alias{eff.2comp.1abs}
\alias{eff.2comp.1abs.ss}
\alias{eff.2comp.iv}
\alias{eff.3comp.1abs}
\alias{pk.1comp.0abs}
\alias{pk.1comp.10abs}
\alias{pk.1comp.1abs}
\alias{pk.1comp.1abs.ss}
\alias{pk.1comp.iv}
\alias{pk.1comp.lag}
\alias{pk.2comp.0abs}
\alias{pk.2comp.1abs}
\alias{pk.2comp.1abs.m}
\alias{pk.2comp.1abs.ss}
\alias{pk.2comp.iv}
\alias{pk.3comp.1abs}
\alias{pk.3comp.1abs.ss}
\alias{pk.3comp.iv}
\alias{pk.3comp.iv.ss}
\alias{pk.pred}
\title{PK predictions using linear equations}
\usage{
pk.pred(doses, t.doses, t.obs, pk.func, parms)

pk.1comp.iv(dose, tob, parms)

pk.1comp.0abs(dose, tob, parms)

pk.1comp.1abs(dose, tob, parms)

pk.1comp.1abs.ss(dose, tob, parms)

pk.1comp.10abs(dose, tob, parms)

pk.1comp.lag(dose, tob, parms)

pk.2comp.iv(dose, tob, parms)

pk.2comp.0abs(dose, tob, parms)

pk.2comp.1abs(dose, tob, parms)

pk.2comp.1abs.m(dose, tob, parms)

pk.2comp.1abs.ss(dose, tob, parms)

pk.3comp.iv(dose, tob, parms)

pk.3comp.iv.ss(dose, tob, parms)

pk.3comp.1abs(dose, tob, parms)

pk.3comp.1abs.ss(dose, tob, parms)

eff.1comp.iv(dose, tob, parms)

eff.1comp.1abs(dose, tob, parms)

eff.1comp.1abs.ss(dose, tob, parms)

eff.2comp.iv(dose, tob, parms)

eff.2comp.1abs(dose, tob, parms)

eff.2comp.1abs.ss(dose, tob, parms)

eff.3comp.1abs(dose, tob, parms)
}
\arguments{
\item{doses}{vector of dose amounts (with pk.pred multiple dose wrapper)}

\item{t.doses}{vector of dose times. Needs to have same length as doses}

\item{t.obs}{vector of observation times relative to time of first dose (with pk.pred multiple dose wrapper)}

\item{pk.func}{function object predicting concentration as a function of time, given    dose and pk parameters}

\item{parms}{vector of pk parameter values. This can but is not required to be a named vector.}

\item{dose}{vector of dose amounts (individual pk.xxx functions)}

\item{tob}{vector of time observations  (individual pk.xxx functions)}

\item{e.func}{function object predicting effects}
}
\value{
PK predictions, PK plots
}
\description{
Collection of linear PK function for 1, 2, and 3 compartment distribution structures and various input functions. Further a collection of effect relationships feeding on concentration from the PK functions. Can be used in isolation but ideally embedded within superposition function \code{pk.pred} and plotting function \code{pkpdPredPlot}
}
\section{Functions}{
\itemize{
\item \code{pk.1comp.iv}: 

\item \code{pk.1comp.0abs}: 

\item \code{pk.1comp.1abs}: 

\item \code{pk.1comp.1abs.ss}: 

\item \code{pk.1comp.10abs}: 

\item \code{pk.1comp.lag}: 

\item \code{pk.2comp.iv}: 

\item \code{pk.2comp.0abs}: 

\item \code{pk.2comp.1abs}: 

\item \code{pk.2comp.1abs.m}: 

\item \code{pk.2comp.1abs.ss}: 

\item \code{pk.3comp.iv}: 

\item \code{pk.3comp.iv.ss}: 

\item \code{pk.3comp.1abs}: 

\item \code{pk.3comp.1abs.ss}: 

\item \code{eff.1comp.iv}: 

\item \code{eff.1comp.1abs}: 

\item \code{eff.1comp.1abs.ss}: 

\item \code{eff.2comp.iv}: 

\item \code{eff.2comp.1abs}: 

\item \code{eff.2comp.1abs.ss}: 

\item \code{eff.3comp.1abs}: 
}}
\examples{
# 1comp elimination, 1st-order absorption
pkPredPlot(doses = rep(100, 7), t.doses = seq(0,24*7,length=7), t.obs = seq(0,24*7)
   , pk.func = pk.1comp.1abs, parms = c(1,25, 0.1, 5,0.5)
)

# 2comp elimination, 0-order absoorption
pkPredPlot(doses = rep(100, 7), t.doses = seq(0,24*7,length=7), t.obs = seq(0,24*7)
   , pk.func = pk.2comp.0abs, parms = c(10,30, 3, 90,10), log = T
)

## demo effect prediction after single and multiple doses
eff.1comp.1abs(dose = 100, tob = seq(0,24*7), parms = c(1,10, 0.25, 0.05))
pk.pred(doses = rep(100, 7), t.doses = seq(0,24*7,length=7), t.obs = seq(0,24*7)
   , pk.func = eff.1comp.iv, parms = c(1,10, 0.25)
)
pkpdPredPlot(doses = rep(100, 7), t.doses = seq(0,24*7,length=7), t.obs = seq(0,24*7)
   , pk.func = pk.1comp.1abs, e.func = eff.1comp.1abs
   , parms = c(1,10, 0.25, 0.05)
)
}

