% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkpdPredPlot.r
\name{pkpdPredPlot}
\alias{pkpdPredPlot}
\title{Title}
\usage{
pkpdPredPlot(doses, t.doses, t.obs, pk.func, e.func, parms,
  output = FALSE)
}
\arguments{
\item{doses}{numeric vector of doses}

\item{t.doses}{numeric vector of time of doses (equal length of \code{doses})}

\item{t.obs}{time opf observations}

\item{pk.func}{pharmacokinetic function to evaluate. See \code{\link{pk.pred}}}

\item{e.func}{effect model function to evaluate. See \code{\link{pk.pred}}}

\item{parms}{numeric vector (may be a named vector but is not required)}

\item{output}{if TRUE the simulation input and outpout is returned as list. If FALSE (default) the plot is created.}
}
\value{
A plot or a list with output
}
\description{
Plots PK-PD predictions based on qP's suite of PK and Effect Model functions
}
\details{
PK-PD superposition of multiple doses
}
\examples{
eff.1comp.iv(dose = 100, tob = seq(0,24*7), parms = c(1,10, 0.25))
pk.pred(doses = rep(100, 7), t.doses = seq(0,24*7,length=7), t.obs = seq(0,24*7),
       pk.func = eff.1comp.iv, parms = c(1,10, 0.25))
pkpdPredPlot(doses = rep(100, 7), t.doses = seq(0,24*7,length=7), t.obs = seq(0,24*7),
            pk.func = pk.1comp.iv, e.func = eff.1comp.iv,
            parms = c(1,10, 0.25))
}
