% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkPredPlot.r
\name{pkPredPlot}
\alias{pkPredPlot}
\title{PK superposition of multiple doses}
\usage{
pkPredPlot(doses, t.doses, t.obs, pk.func, parms, log = FALSE, output = FALSE)
}
\arguments{
\item{doses}{numeric vector of doses}

\item{t.doses}{numeric vector of time of doses (equal length of \code{doses})}

\item{t.obs}{time opf observations}

\item{pk.func}{pharmcokinetic function to evaluate. See \code{\link{pk.pred}}}

\item{parms}{numeric vector (may be a named vector but is not required) with}

\item{log}{logical indicating to log the Y axis}

\item{output}{if TRUE the simulation input and outpout is returned as list. If FALSE (default) the plot is created.}
}
\value{
PK plot
}
\description{
Plots PK predictions based on qP's suite of PK functions
}
\examples{
# 1comp elimination, 1st-order absorption
pkPredPlot(doses = rep(100, 7), t.doses = seq(0,24*7,length = 7), t.obs = seq(0,24*7)
   , pk.func = pk.1comp.1abs, parms = c(1,25, 0.1, 5,0.5)
)

# 2comp elimination, 0-order absoorption
pkPredPlot(doses = rep(100, 7), t.doses = seq(0,24*7,length = 7), t.obs = seq(0,24*7)
   , pk.func = pk.2comp.0abs, parms = c(10,30, 3, 90,10), log = TRUE
)

## demo effect prediction after single and multiple doses
eff.1comp.1abs(dose = 100, tob = seq(0,24*7), parms = c(1,10, 0.25, 0.05))
pk.pred(doses = rep(100, 7), t.doses = seq(0,24*7,length = 7), t.obs = seq(0,24*7)
   , pk.func = eff.1comp.iv, parms = c(1,10, 0.25)
)
pkpdPredPlot(doses = rep(100, 7), t.doses = seq(0,24*7,length = 7), t.obs = seq(0,24*7)
   , pk.func = pk.1comp.1abs, e.func = eff.1comp.1abs
   , parms = c(1,10, 0.25, 0.05)
)
}
\seealso{
\code{\link{pk.pred}}, \code{\link{pk.1comp.iv}}, \code{\link{pk.1comp.0abs}}, \code{\link{pk.1comp.1abs}}, \code{\link{pk.1comp.1abs.ss}}, \code{\link{pk.1comp.10abs}}, \code{\link{pk.1comp.lag}}, \code{\link{pk.2comp.iv}}, \code{\link{pk.2comp.0abs}}, \code{\link{pk.2comp.1abs}}, \code{\link{pk.2comp.1abs.m}}, \code{\link{pk.2comp.1abs.ss}}, \code{\link{pk.3comp.iv}}, \code{\link{pk.3comp.iv.ss}}, \code{\link{pk.3comp.1abs}}, \code{\link{pk.3comp.1abs.ss}}, \code{\link{eff.1comp.iv}}, \code{\link{eff.1comp.1abs}}, \code{\link{eff.1comp.1abs.ss}}, \code{\link{eff.2comp.iv}}, \code{\link{eff.2comp.1abs}}, \code{\link{eff.2comp.1abs.ss}}, \code{\link{eff.3comp.1abs}}
}
