% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.process.coveffects.r
\name{plot.covEffects}
\alias{plot.covEffects}
\title{Plot a covariate effects object}
\usage{
\method{plot}{covEffects}(obj, idx = names(obj$plot))
}
\arguments{
\item{idx}{Names of the plots to print. Defaults to all plots.}

\item{a}{covEffects object from nm.process.coveffects}
}
\description{
Plot a covariate effects object
}
