% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tornado.r
\name{plot.tornado}
\alias{plot.tornado}
\title{Tornado plots}
\usage{
\method{plot}{tornado}(
  x,
  ...,
  main = "",
  xlim = range(c(torn.ds$val.hi, torn.ds$val.lo)),
  xlab = "Output",
  bar.width = 0.75,
  base.round = 0,
  cex.main = 1.5,
  cex.xaxis = cex.main * 0.5,
  cex.yaxis = cex.main * 0.65,
  cex.barlab = cex.main * 0.55,
  cex.baslab = cex.main * 0.5,
  bar.col = "#0A41A5",
  show.base = TRUE,
  adj.ylabs = 0
)
}
\arguments{
\item{x}{Dataset containing sensitivity analysis results to be plotted. Column names should include: \code{variable} - names of variable (will be plotted as y axis labels), \code{val.lo} - low value of objective when variable is swung to an extreme, \code{val.base}    base value of objective when all variables are at base, \code{val.hi} - high value of the objective when variable is swung to an extreme, \code{lab.lo} -    label associated with low, \code{lab.base} -    label associated with base, \code{lab.hi}    label associated with high}

\item{\dots}{ignored}

\item{main}{graphical parameter}

\item{xlim}{graphical parameter}

\item{xlab}{graphical parameter}

\item{bar.width}{graphical parameter}

\item{base.round}{graphical parameter}

\item{cex.main}{graphical parameter}

\item{cex.xaxis}{graphical parameter}

\item{cex.yaxis}{graphical parameter}

\item{cex.barlab}{graphical parameter}

\item{cex.baslab}{graphical parameter}

\item{bar.col}{graphical parameter}

\item{show.base}{graphical parameter}

\item{adj.ylabs}{graphical parameter}
}
\value{
A tornado plot
}
\description{
Create a tornado plot given a table of sensitivity analysis results
}
\examples{
torn.ds = as.tornado(read.csv(
  file.path(getOption("qpExampleDir"), "../Excel/tornado.ds.csv")
, stringsAsFactors = FALSE))
plot(torn.ds
             , xlim = c(0,20)
             , xlab = "Dollars"
             , main = "My first tornado"
             , cex.main = 2
             , bar.width =.50
             , base.round = 1
             , show.base = FALSE
)
}
\seealso{
Other tornado: 
\code{\link{as.tornado.data.frame}()},
\code{\link{as.tornado}()}
}
\concept{tornado}
