% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/png.slide.full.r
\name{png.slide.full}
\alias{png.slide.full}
\title{PNG graph for PowerPoint generation (full size)}
\usage{
png.slide.full(filename = "C:/Test/test.png", w = 10, h = 7.5,
  units = "in", psize = 10, bg = "transparent", res = 600, ...)
}
\arguments{
\item{filename}{filename including path. To follow qP workflow standards should read the internal relative reference directory "../WorkAre/output/graphs"}

\item{w}{width in \code{units}}

\item{h}{height in \code{units}}

\item{units}{units ('in', 'cm', or 'pt')}

\item{psize}{pointsize in \code{units}}

\item{bg}{background color}

\item{res}{resolution of PNG graph. High values create top quality graph at the expense of large file size.}

\item{...}{any additional parameters to be passed on the \code{png()}}
}
\value{
A graph written to file in .png format that does not have to be edited when pasted into MS PowerPoint(r).
}
\description{
A wrapper function for a call to png(), which opens a plot device (i.e. enables you to "print" a graphic to a file), and provides defaults for various parameters of that graphic suitable for a slide in a presentation. The plot area will over the full slide dimension and does not have to be resized or repositioned after pasting. Note that the plot must be generated and then \code{dev.off()} must be called to close the plot device (i.e. close the output graphic file)
}
\examples{
png.slide.full(filename="../WorkArea/output/graphs/myPNGslidefullexample.png")
set.seed(1234567)
plot(x=1:25,y=rnorm(25,50,5))
dev.off()
}
\seealso{
\code{\link{png.slide.semi}}, \code{\link{png.report}}
}

