% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.process.coveffects.r
\name{print.covInfo}
\alias{print.covInfo}
\title{Print a covInfo S3 object}
\usage{
\method{print}{covInfo}(x)
}
\arguments{
\item{x}{A covInfo object}
}
\value{
A covInfo object
}
\description{
Prints a covInfo object.
}
\note{
This function is used with \code{nm.process.coveffects}
}
