% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.parTable.r
\name{process.parTable}
\alias{process.parTable}
\title{Format parameter estimate table}
\usage{
process.parTable(
  nm,
  index = 1,
  ci = 0.95,
  sig = 3,
  transformations,
  Descriptor,
  plain = FALSE,
  missing.format = ".....",
  remove.fixed.sigma = FALSE
)
}
\arguments{
\item{nm}{output from \code{nm.params.table}}

\item{index}{which estimation (only relevant in case of multiple \code{$EST})}

\item{ci}{confidence interval for 95\% CI}

\item{sig}{number of significant digits in output}

\item{transformations}{a list with elements log or logit, that, in turn, are numeric vectors representating the THETA numbers that are logged or logit-transformed, respectively.}

\item{Descriptor}{vector of descriptor (alias) names for numbered parameter estimates. For example c("CL","V1","SD(CL)","cor(CL,V1)","SD(V1)","prop. residual var.")}

\item{plain}{logical (defaults to F) indicating if LaTeX formatting of the table should be skipped, allowing the result to be easily available for Microsoft products.}

\item{missing.format}{what to replace non-estimated / fixed parameter estimates, or}

\item{remove.fixed.sigma}{logical option to remove sigma estimates from the table}
}
\value{
Either a LaTeX preformatted parameter estimate table (default) or a fully processed parameter estimate table without LaTeX preformatting (for inclusion into Microsoft products).
}
\description{
Process parameter table for inclusion into LaTeX, or MS Excel or Word
}
\examples{
process.parTable(
   nm = nm.params.table(run = "example1", path = getOption("qpExampleDir"), runIndex = 1)
 , transformations = list(log=c(1:12))
 , missing.format = "...."
)
process.parTable(
   nm = nm.params.table(run = "example1", path = getOption("qpExampleDir"))
 , transformations = list(logit=c(7,9))
 , plain = TRUE
 , missing.format = "...."
)
}
\seealso{
\code{\link{nm.params.table}}, \code{\link{process.runrec}}
}
