% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qP.colors.r
\docType{data}
\name{qp.colors.sorted}
\alias{qp.colors.sorted}
\title{qP Colors Sorted}
\format{
An object of class \code{character} of length 22.
}
\usage{
qp.colors.sorted
}
\description{
A sorted version of qP colors
}
\keyword{datasets}
