% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpToolkit.R
\docType{package}
\name{qpToolkit}
\alias{qpToolkit}
\alias{qpToolkit-package}
\title{qpToolkit: Tools for population PK/PD modeling, visualization, and reporting.}
\description{
The qpToolkit package provides different categories of important functions:
\itemize{
\item{R modeling functions}
\item{NONMEM post-processing}
\item{Data tables}
\item{R qualification}}
}
\section{qpToolkit functions}{

Watch here for forthcoming summaries of the qpToolkit functions.
}

