% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqGamma.r
\name{qqGamma}
\alias{qqGamma}
\title{Plot QQ Gamma distribution}
\usage{
qqGamma(
  x,
  ylab = deparse(substitute(x)),
  xlab = "Theoretical Quantiles",
  main = "Gamma Distribution QQ Plot",
  ...
)
}
\arguments{
\item{x}{numeric vector suspected to be Gamma distributed}

\item{ylab}{y label}

\item{xlab}{x label}

\item{main}{main title of the plot}

\item{\dots}{passed to qqplot}
}
\value{
character vector with the NONMEM output. Each element i the vector represents a line in the NONMEM output file
}
\description{
Plot qq-plot for gamma distributed variable
}
\examples{
qqGamma(rgamma(100,shape = 0.1,scale = 0.2))

}
