% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/R.installation.validation.r
\name{R.installation.validation}
\alias{R.installation.validation}
\title{R Installation Validation}
\usage{
R.installation.validation()
}
\value{
Information
}
\description{
R Installation Validation
}

