% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bootstrap.r
\name{read.bootstrap}
\alias{read.bootstrap}
\title{Documentation stub}
\usage{
read.bootstrap(
  path = getOption("nmDir"),
  filename = "raw_results1.csv",
  structure.filename = "raw_results_structure",
  parlist = c("ofv", "theta", "omega")
)
}
\arguments{
\item{path}{Directory where bootstrap summary files reside}

\item{filename}{default Filename of the bootstrap results summary file created by PsN bootstrap}

\item{structure.filename}{Filename of structure file created by PsN bootstrap}

\item{parlist}{Define what elements are needed from the bootstrap?}
}
\value{
A list with data frame withy summary boostrap results
}
\description{
Reads summary files of a completed NONMEM bootstrap
}
\examples{
\dontrun{
myBoot = read.bootstrap(path = getOption("qpExampleDir")
  , filename = "raw_results_bs4011.csv"
  , structure.filename = "raw_results_structure"
  )
bootstrap.ParTab(myBoot, idx = list(theta=1:13,omega=1:7,sigma=1))
names(myBoot)
str(myBoot$bootstrap)
myBoot$structure
}

}
