% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.ext.r
\name{read.ext}
\alias{read.ext}
\title{Parse the information from the .ext file}
\usage{
read.ext(run, path = getOption("nmDir"), file.ext = ".ext", quiet = T)
}
\arguments{
\item{run}{run rootname (e.g. \code{run1})}

\item{path}{directory where \code{run} resides}

\item{quiet}{if TRUE no message will be returned.
#'}

\item{extension}{file extension of the .ext file}
}
\value{
A list with data.frames (as many as there where \code{$EST} calls in the NONMEM control stream) containing the .ext output
}
\description{
Read and parse the  information embedded in the .ext file from a completed NONMEM run. In case of multiple \code{$EST} calls, these will be parsed one by one and included with their respective \code{$EST} names into the resulting list
}
\examples{
read.ext(run = "example1", path = getOption("qpExampleDir"))
## now parse an output where multiple $EST calls have been performed
read.ext(run = "example1", path = getOption("qpExampleDir"), file.ext = ".mext")
## a large one
read.ext(run = "example1", path = getOption("qpExampleDir"))
}
\seealso{
\code{\link{read.out}}
}

