% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.out.r
\name{read.out}
\alias{read.out}
\title{Scan NONMEM output file}
\usage{
read.out(run, path = getOption("nmDir"), file.ext = ".lst", quiet = TRUE)
}
\arguments{
\item{run}{run rootname (e.g. \code{run1})}

\item{path}{directory where \code{run} resides}

\item{file.ext}{file extension of the NONMEM output file (.lst)}

\item{quiet}{if TRUE no message will be returned.}
}
\value{
character vector with the NONMEM output. Each element i the vector represents a line in the NONMEM output file
}
\description{
Scan and parse the NONMEM output (lst file) as scanned lines. This function was tested for NONMEM v7.2 and v7.3.
}
\examples{
head(read.out(path = getOption("qpExampleDir"), run = "example1", file.ext = ".lst"))
}
\seealso{
\code{\link{read.ext}} \code{\link{get.ofv}}, \code{\link{get.dOFV}}, \code{\link{get.eigen}}, \code{\link{get.shrinkage}}, \code{\link{get.description}}
}
