% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.runrec.r
\name{read.runrec}
\alias{read.runrec}
\title{Read the runrecord}
\usage{
read.runrec(filename, path = getOption("nmDir"))
}
\arguments{
\item{filename}{filename, e.g. "AAinfo.txt"}

\item{path}{diretory where the runrecord resides}
}
\value{
Data.frame with all information extracted from the runrecord.
}
\description{
Read the run record into readable table. This function assumes runrecord has been run with PsN option -maxlvl = 0. It is to be used in conjunction with process.runrec
}
\examples{
rr = read.runrec(filename = "AAruninfo.txt"
, path = system.file(package = 'qpToolkit','NONMEM')
)
rr
process.runrec(rr)
}
\seealso{
\code{\link{process.runrec}}
}
