% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.vpc.r
\name{read.vpc}
\alias{read.vpc}
\title{Create covariate plot and data}
\usage{
read.vpc(path = "./", result = result, tab = tab)
}
\arguments{
\item{path}{directory where VPC output resides}

\item{result}{filename of the 'vpc_results.csv" file}

\item{tab}{filename of the 'vpctab' file}
}
\value{
list of data frames used to plot a VPC
}
\description{
An alternative VPC plotting routine.
}
\note{
This function is used by nm.read.vpc and is intended for that purpose only; running this function alone will not help the average R programmer. Please note that qPharmetra's default VPC plot tool is xpose.VPC. Use this tool in case one needs more flexibility to post-process results before plotting, and when stratifcation by more than one variable is needed.
}
\seealso{
\code{\link{nm.read.vpc}}, \code{\link{ggvpc_standard}}, \code{\link{ggvpc_xpose}}
}
