% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.tex.files.r
\name{remove.tex.files}
\alias{remove.tex.files}
\title{Remove intermediate LaTeX Files}
\usage{
remove.tex.files(name = "qPharmetra", path)
}
\arguments{
\item{name}{rootname of the LaTeX files.}

\item{path}{the report folder where intermediate LaTeX files reside}
}
\value{
Intermediate LaTeX files are deleted - no visible R result.
}
\description{
Strips report area of latent tex files that should ideally be recreated every latex run
}
