% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeTrailing.r
\name{removeTrailing}
\alias{removeTrailing}
\title{Remove trailing items from vector}
\usage{
removeTrailing(x, val = 0)
}
\arguments{
\item{x}{numeric vector}

\item{val}{the trailing value to be removed}
}
\value{
Numeric vector \code{x} without trailing \code{val}
}
\description{
Removes items from the end of a numeric vector that are equal to the value supplied
}
\examples{
conc = c(1:30,0,0,0,0)
conc
removeTrailing(conc, val = 0)
}
