% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeath.nth.r
\name{repeat.nth}
\alias{repeat.nth}
\title{Hide repetition of character strings}
\usage{
repeat.nth(x, replacement = "", before = "", after = "", pos = 1)
}
\arguments{
\item{x}{numeric or character}

\item{replacement}{character string to replace the repeated entries (defaults to the empty "")}

\item{before}{prefix (optional)}

\item{after}{suffix (optional)}

\item{pos}{position within the repetition the non-empty character string should appear}
}
\value{
Character vector
}
\description{
Creates a vector in which text is repeated every "nth" item in the result, and in which that text is concatenated with a numeric integer indicating which occurence it represents. Useful for making good-looking, not-too busy Appendix tables of for example observed concentrations  with multiple samples taken at various dose levels and/or various dosing cycles.
}
\examples{
## simplest form
repeat.nth(c(1,1,1,1,2,2,2))

## adding a prefix
repeat.nth(c(1,1,1,1,2,2,2), before = "CYCLE", pos = 1)
repeat.nth(c(1,1,1,1,2,2,2), before = "CYCLE", pos = 2)
repeat.nth(c(1,1,1,1,2,2,2), before = "CYCLE", pos = 3)

## addding a suffix
repeat.nth(c(1,1,1,1,2,2,2), after = "CYCLE", pos = 1)
repeat.nth(c(1,1,1,1,2,2,2), after = "CYCLE", pos = 2)
repeat.nth(c(1,1,1,1,2,2,2), after = "CYCLE", pos = 3)

## both
repeat.nth(c(1,1,1,1,2,2,2), before = "BI", after = "CYCLE", pos = 1)
}
