% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgenpois.r
\name{rgenpois}
\alias{rgenpois}
\title{Sample from Generalized Poisson distribution}
\usage{
rgenpois(lambda = lambda, disp = disp, n = 1, scan.size = 100)
}
\arguments{
\item{lambda}{lambda of the Generalized Poisson distribution}

\item{disp}{dispersion of the Generalized Poisson distribution. Takes the value of 0 for a Poisson distribution, and becomes negative for under, and positive for over dispersed data.}

\item{n}{sample size}

\item{scan.size}{the upper limit of the distribution. This could be set higher but due to exponentation in the likelihood function shgould not exceed}
}
\value{
n samples from the Generalized Poisson distribution (numeric non-negative integers)
}
\description{
Sample from Generalized Poisson distribution
}
\examples{
genPois.samples = sapply(seq(0.3,-1.1,-0.2), function (DISP)
{
  rgenpois(n = 10000, lambda = 22, disp = DISP)
})

# plot across dispersion
plot(range(genPois.samples), c(0,.25), type = "n", xlim = c(0,100), ylab = "Frequency")
for(i in 1 : ncol(genPois.samples)){
  gpDen = density(genPois.samples[,i], bw = 1)
  lines(gpDen)
  text(mean(genPois.samples[,i])
       , max(gpDen$y)+0.015
       , round(mean(genPois.samples[,i]),1)
       , cex = 0.5
  )
}
lines(density(rpois(10000, 11),bw = 1), col = qp.green, lwd = 3)
text(3, 0.03, round(mean(rpois(10000, 11)), 1), col = qp.green)
}
