% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorm.by.id.r
\name{rnorm.by.id}
\alias{rnorm.by.id}
\title{Group-wise sampling from normal distribution}
\usage{
rnorm.by.id(id, mean = 0, sd = 1)
}
\arguments{
\item{id}{vector with group information}

\item{mean}{mode of the normal distribution}

\item{sd}{standard deviation of the normal distribution}
}
\value{
samples of a normal distribution, one for each level of \code{group}
}
\description{
Populates 1 sample from normal distribution for each group (id or subject, or trial, or other)
}
\examples{
df = example.pkpdData()
sam = rnorm.by.id(df$id, mean = 0, sd = 1)
tapply(sam,df$id, unique) 
}
\seealso{
\code{\link{sample.by.id}}
}
