% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runrec.R
\name{runrec}
\alias{runrec}
\title{Construct a Run Record}
\usage{
runrec(x, prefix = "run", suffix = "", ...)
}
\arguments{
\item{x}{vector of integers representing models}

\item{prefix}{character string to prepend to x}

\item{suffix}{character string to append to x}

\item{...}{ignored}

\item{mods}{model names constructed from x, prefix, and suffix by default}
}
\value{
decorated data.frame
}
\description{
Constructs a psn-style run record.
}
\keyword{internal}
