% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.by.id.r
\name{sample.by.id}
\alias{sample.by.id}
\title{Group-wise sampling from prespecified distribution}
\usage{
sample.by.id(id, samples, replace = TRUE)
}
\arguments{
\item{id}{vector with group information}

\item{samples}{distribution to sample from}

\item{replace}{if TRUE (default) sampling will be done with replacement}
}
\value{
samples of the prespecified distribution, one for each level of \code{group}
}
\description{
Populates 1 sample from prespecified distribution for each group (id or subject, or trial, or other)
}
\examples{
my.ids = rep(1:5,each=3)
my.ids

set.seed(123456)
sample.by.id(my.ids, samples=rgamma(1000,1))

my.ids = rep(1:5,times=3)
my.ids
sample.by.id(my.ids, samples=rgamma(1000,1))

#one unique value per id
sample.by.id(unique(my.ids), samples=rgamma(1000,1))
tapply(sample.by.id(my.ids, samples = rgamma(1000,1)),my.ids, unique)
}
\seealso{
\code{\link{rnorm.by.id}}
}
