% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sgeomean.cl.normal.r
\name{sgeomean.cl.normal}
\alias{sgeomean.cl.normal}
\title{Summary of geometric mean and 95\% CI}
\usage{
sgeomean.cl.normal(x, conf.int = 0.95, log.base = exp(1), na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{conf.int}{confidence interval, default is 0.95, for 95\%)}

\item{log.base}{base of log, defaults to natural log}

\item{na.rm}{logical determining if NA values need to be omitted before the operation}
}
\value{
Vector of three values, representing the geometric mean, and the lower and  upper component of the confidence interval.
}
\description{
Make a summary of the geometric mean and 95\% CI of the input
}
\examples{
test.vector = rlnorm(1000,log(100),0.5)
hist(test.vector)
smean.cl.normal(test.vector) ## arithmetic mean and 95\% confidence intervals
sgeomean.cl.normal(test.vector) ## geometic mean and 95\% confidence intervals
}

