% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shuffle.list.r
\name{shuffle.list}
\alias{shuffle.list}
\title{Shuffle/reorder lists}
\usage{
shuffle.list(x)
}
\arguments{
\item{x}{list}
}
\value{
A multi-level list
}
\description{
Shuffles a multi-level list
}
\note{
Is called by qP function nm.extract.xml
}
\examples{
test = list(A=list(a=1,b=2,c=3), B = list(a=2,b=1,c=0))
shuffle.list(test)
}

