% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superPosition.r
\name{superPosition}
\alias{superPosition}
\title{Superposition of concentration-time profiles}
\usage{
superPosition(val, time, tau)
}
\arguments{
\item{val}{numeric vector}

\item{time}{sorted numeric vector (positive values only)}

\item{tau}{time interval between two doses}
}
\value{
numeric vector with super positioned results
}
\description{
Superposition of concentration-time profiles.
}
\examples{
pPK = superPosition(15*exp(-0.15*seq(0,84)), time = seq(0,84), tau = 8)
plot(seq(0,84), pPK, type = 'l')
}
