% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.r
\name{swap}
\alias{swap}
\title{Exchange Existing set of Values for another set of Values}
\usage{
swap(x, uniques, newUniques)
}
\arguments{
\item{x}{Vector to apply the exchange over}

\item{uniques}{(Sorted) vector of unique OLD values}

\item{newUniques}{(Sorted) vector of unique NEW values}
}
\value{
Swapped vector of x
}
\description{
Exchange Existing set of Values for another set of Values
}
\examples{
my.vect = c(1,2,2,2,4,4,4,3,3,3,3)
swap(my.vect,c(1,2,3,4),c(11,22,33,44))
}
