% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.meta.analysis.summary.r
\name{table.meta.analysis.summary}
\alias{table.meta.analysis.summary}
\title{Summarize meta-analysis dataset}
\usage{
table.meta.analysis.summary(data, cov.cols = NULL, cov.cat = NULL)
}
\arguments{
\item{data}{meta.analysis dataset}

\item{cov.cols}{character vector with covariate columns}

\item{cov.cat}{character vector indicating whether each covariate is continuous or categorical}
}
\value{
Summarized meta-analysis dataset
}
\description{
Creates a summary of a meta-analysis dataset. The dataset needs specific column names (see Notes).
}
\note{
Standard column names are
reference	reference number or label to uniquely identify each public literature article/source
	study		study identifier or label
	drug		drug name in a given arm
	dose		dose in a given arm (could be total daily, or some other summary figure)
	time		observation time
	n		sample size at each observation time
	Each unique study+drug+dose = one arm

	If there are multiple doses of a drug in a study, computes the AVERAGE starting size
Covariate columns (both continous and categorical) can be specified in the "cov.cols" argument (a vector).
Whether each covariate is continous or categorical is specified in the "cov.cat" argument.
}
\examples{
meta.analysis.ds = read.csv(file.path(getOption("qpExampleDir")
 , "../Excel/meta.analysis.ds.csv"),stringsAsFactors = FALSE)
table.meta.analysis.summary(meta.analysis.ds
 , cov.cols=c("age","fem.pct","dis.score.bl","drug.ir")
 , cov.cat=c(FALSE,FALSE,FALSE,TRUE)
)
}
