% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabStats.r
\name{tabStats}
\alias{tabStats}
\title{Table Statistics}
\usage{
tabStats(x, BY, nSignif = 3, conFunc1 = conDataFun1,
  conFunc2 = conDataFun2, catFunc = catDataFun,
  ndigits.categorical = 1, parName)
}
\arguments{
\item{x}{numeric vector to summarize}

\item{BY}{stratification argument}

\item{nSignif}{number of significant digits in continuous data summaries}

\item{conFunc1}{defaults to conDataFun1}

\item{conFunc2}{defaults to conDataFun2}

\item{catFunc}{defaults to catDataFun}

\item{ndigits.categorical}{passed to catFunc}

\item{parName}{is automatically populated but the parameter name can be specified here by the user.}
}
\value{
Appropriate statistic for x
}
\description{
Determines if variable is continuous or categorical and returns the appropriate summary statistics.
}
\note{
This function is primarily used for demographics tables
}
\examples{
 
pkpdData = example.pkpdData()
ok = duplicated(pkpdData$id) == FALSE
tabStats(x=pkpdData$race[ok], BY=list(dose = pkpdData$dose[ok]))
tabStats(x=pkpdData$race[ok], BY=list(dose = pkpdData$dose[ok]), parName = "Race", 
ndigits.categorical=0)
tabStats(pkpdData$sex[ok], list(dose = pkpdData$dose[ok]))
tabStats(x=pkpdData$wt, BY = list(dose = pkpdData$dose))
tabStats(x=pkpdData$bmi, BY = list(dose = pkpdData$dose))
}
\seealso{
\code{\link{conDataFun1}},  \code{\link{conDataFun2}},  \code{\link{conDataFun3}},  \code{\link{catDataFun}},  \code{\link{tabSummarize}}
}
