% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform.log10.r
\name{transform.log10}
\alias{transform.log10}
\title{Log10 transformation}
\usage{
\method{transform}{log10}(x, log)
}
\arguments{
\item{x}{Value to log with base 10}

\item{log}{If TRUE or any numeric value the log base 10 of x will be taken}
}
\value{
log10(x)
}
\description{
Takes the log10 of the input in case argument log is T or a numeric value
}
\examples{
transform.log10(c(1,10,100), log = T)
transform.log10(c(2,3,200,1000,33330), 2)
}

