% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trimSpace.r
\name{trimSpace}
\alias{trimSpace}
\title{Trimming spaces}
\usage{
trimSpace(x)
}
\arguments{
\item{x}{character string}
}
\value{
A character string with any spaces at its beginning or end trimmed off
}
\description{
Remove space from beginning and/or end of a string (trim).
}
\examples{
trimSpace(" Hello ")
trimSpace(" Hello there ")
}

