% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nm.create.control.r
\name{uncertainParams}
\alias{uncertainParams}
\title{Sample parameters across uncertainty}
\usage{
uncertainParams(out)
}
\arguments{
\item{out}{output from nm.extract.xml}
}
\value{
single draw from quasi-posterior parameter distribution
}
\description{
Sample parameters across uncertainty
}

