% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unitDensity.r
\name{unitDensity}
\alias{unitDensity}
\title{Density plot normalized to unity}
\usage{
unitDensity(x, ...)
}
\arguments{
\item{x}{vector of numeric values to pass on to a \code{density}}

\item{...}{any arguments to be passed on to \code{density default}}
}
\value{
list of \code{x} and \code{y} with x and y information to generate a plot
}
\description{
Create a density of inout and normalizes the maximum to be 1
}
\examples{
standardNormal = unitDensity(qnorm(seq(0.0001,0.9999,length=1000),0,1))
plot(standardNormal$x, standardNormal$y, col = qp.blue, type = 'l')
}

