% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitHistogram.r
\name{unitHistogram}
\alias{unitHistogram}
\title{Histogram normalized to unity}
\usage{
unitHistogram(x, myCuts = NULL, lwd.bar = 4, offset = 0.4,
  hist.type = "box", col.hist = "#D7D7D7", xlim = NULL,
  ylab = NULL, ...)
}
\arguments{
\item{x}{numeric vector}

\item{myCuts}{alternative cut points (optional)}

\item{lwd.bar}{line width of the histogram bars}

\item{offset}{width between the bars}

\item{hist.type}{histogram type, can be 'box' or 'lines'}

\item{col.hist}{color of histograms}

\item{xlim}{limits of x coordinate}

\item{ylab}{label of the frequency (y label)}

\item{\dots}{passed to plot()}
}
\value{
A histogram
}
\description{
Plots a histogram of data provided in which the height of the frequency are normalized.
}
\examples{
set.seed(1234567)
myvals = rnorm(10000,50,5)
unitHistogram(myvals, xlab = "My Values", cex.lab = 1.25)
}
