% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2df.r
\name{vec2df}
\alias{vec2df}
\title{Turn vector into data.frame}
\usage{
vec2df(x, n = 1)
}
\arguments{
\item{x}{vector}

\item{n}{number of times the vector needs to be repeated}
}
\value{
data.frame.
}
\description{
Turns a vector into a data frame with n repeated instances of the vector.
}
\note{
This function is used by \code{nm.covplot}
}
\examples{
vec2df(x=1:10,n=1)
vec2df(x=1:10,n=2)
}
