% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichNumeric.r
\name{whichNumeric}
\alias{whichNumeric}
\title{Which entries are numeric?}
\usage{
whichNumeric(x)
}
\arguments{
\item{x}{a vector}
}
\value{
position in the character of the elements that have numbers that can be extracted
}
\description{
Identify which members of a character vector will translate to numeric mode
}
\examples{
test = c("123xyc","A","<LOQ","547.39", "<0.15")
whichNumeric(test)
extract.number(test[whichNumeric(test)])
}
