% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.Extract.NM.Matrix.r
\name{xml.Extract.NM.Matrix}
\alias{xml.Extract.NM.Matrix}
\title{XML parser for matrices}
\usage{
xml.Extract.NM.Matrix(y)
}
\arguments{
\item{y}{XML children}
}
\value{
A matrix
}
\description{
Extract XML content into a Numeric Matrix
}
\examples{
\dontrun{
library(XML)
run = "example1"
filename = "C:/nm73g64/examples/example1.xml"
top = xmlRoot(xmlTreeParse(filename))# names(top)
TMP = top[["nonmem"]][["problem"]]
TMP = TMP[grepl("estimation", names(TMP))]
len = length(TMP)
## get estimation titles
est.methods = sapply(1:length(TMP), function(x,TMP)
  xmlValue(TMP[x][["estimation"]][["estimation_title"]]), TMP = TMP)
names(TMP) = est.methods
tmp = TMP[[1]]

xml.Extract.NM.Value(asXMLNode(tmp[["theta"]]), what = "theta")
xml.Extract.NM.Value(asXMLNode(tmp[["thetase"]]), what = "theta")
names(tmp)
omega = lapply(asXMLNode(tmp[["omega"]])$children, xml.Extract.NM.Matrix)

make.symmetric(unlist(omega))
}
}
\seealso{
\code{\link{nm.extract.xml}}, \code{\link{make.symmetric}}
}
