% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.Extract.NM.Value.r
\name{xml.Extract.NM.Value}
\alias{xml.Extract.NM.Value}
\title{XML parser for vectors}
\usage{
xml.Extract.NM.Value(x, what = "theta")
}
\arguments{
\item{x}{The 'theta' element from NONMEM output in XML format}

\item{what}{character string, like 'theta' or 'thetase'}
}
\value{
Numeric vector
}
\description{
Extract XML content into a numeric vector
}
\examples{
\dontrun{
library(XML)
run = "example1"
filename = "C:/nm73g64/examples/example1.xml"
top = xmlRoot(xmlTreeParse(filename))# names(top)
TMP = top[["nonmem"]][["problem"]]
TMP = TMP[grepl("estimation", names(TMP))]
len = length(TMP)
## get estimation titles
est.methods = sapply(1:length(TMP), function(x,TMP)
  xmlValue(TMP[x][["estimation"]][["estimation_title"]]), TMP = TMP)
names(TMP) = est.methods
tmp = TMP[[1]]

xml.Extract.NM.Value(asXMLNode(tmp[["theta"]]), what = "theta")
xml.Extract.NM.Value(asXMLNode(tmp[["thetase"]]), what = "theta")
}
}
\seealso{
\code{\link{nm.extract.xml}}
}
