% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xscale.components.log10.r
\name{xscale.components.log10.3}
\alias{xscale.components.log10.3}
\title{Labels 10-logged x axes appropriately in lattice (ticks at powers of 1 and 3)}
\usage{
xscale.components.log10.3(lim, ...)
}
\arguments{
\item{lim}{internal required argument.}

\item{\dots}{passed to xscale.components.default()}
}
\value{
Nothing. It's a function to be called inside a lattice call
}
\description{
Labels 10-logged x axes appropriately in lattice (ticks at powers of 1 and 3)
}
\note{
Not to be called by user.
}
\examples{
library(lattice)
set.seed(1234)
datf = data.frame(TIME = exp(rnorm(3000, sd=0.6)), DV = seq(3000))

xyplot(DV ~ TIME
       , data = datf
       , scales = list(x = list(log = 10))
       , xscale.components = xscale.components.log10.3
)
}
\seealso{
Other scale.components: \code{\link{xscale.components.log10}},
  \code{\link{yscale.components.log10.3}},
  \code{\link{yscale.components.log10}}
}
\concept{scale.components}
