% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xscale.components.log10.r
\name{xscale.components.log10}
\alias{xscale.components.log10}
\title{Nice log10 y axes labels}
\usage{
xscale.components.log10(lim, ...)
}
\arguments{
\item{lim}{internal required argument.}
}
\value{
Nothing. It's a function to be called inside a lattice call
}
\description{
Labels 10-logged x axes appropriately in lattice
}
\note{
Not to be called by user.
}
\examples{
set.seed(1234)
datf = data.frame(DV = exp(rlnorm(1000, sdlog=0.5)), TIME = seq(1000))

# labels at one log10 unit intervals
xyplot(DV ~ TIME
       , data = datf
       , scales = list(log = 10)
       , yscale.components = yscale.components.log10
       , xscale.components = xscale.components.log10
)

# labels at half log10 unit intervals
xyplot(DV ~ TIME
       , data = datf
       , scales = list(log = 10)
       , yscale.components = yscale.components.log10.3
       , xscale.components = xscale.components.log10.3
)
}
\seealso{
\code{\link{yscale.components.log10}}, \code{\link{yscale.components.log10.3}}
}
