% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yscale.components.log10.r
\name{yscale.components.log10.3}
\alias{yscale.components.log10.3}
\title{Labels 10-logged y axes appropriately in lattice (ticks at powers of 1 and 3)}
\usage{
yscale.components.log10.3(lim, ...)
}
\arguments{
\item{lim}{internal required argument.}

\item{\dots}{passed to yscale.components.default()}
}
\value{
Nothing. It's a function to be called inside a lattice call
}
\description{
Labels 10-logged y axes appropriately in lattice (ticks at powers of 1 and 3)
}
\note{
Not to be called by user.
}
\seealso{
Other scale.components: 
\code{\link{xscale.components.log10.3}()},
\code{\link{xscale.components.log10}()},
\code{\link{yscale.components.log10}()}
}
\concept{scale.components}
