% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap.parTab.r
\name{bootstrap.ParTab}
\alias{bootstrap.ParTab}
\title{Create parameter table from NONMEM bootsrap}
\usage{
bootstrap.ParTab(bootstrap, idx = list(theta = 1, omega = 1, sigma = 1),
  probs = 0.95, central = median, nsig = 3)
}
\arguments{
\item{bootstrap}{A bootstrap object (list) created by read.bootstrap.}

\item{idx}{list of which parameters to include.}

\item{probs}{The confidence interval coverage.}

\item{central}{function for central tendency of distribution}

\item{nsig}{number of significant digits in output}
}
\value{
A data.frame with the bootstrap estimates.
}
\description{
Create parameter table from NONMEM bootsrap
}
\examples{
myBoot = read.bootstrap(path = getOption("qpExampleDir"),filename = "bootstrap/raw_results_bs4011.csv",structure.filename = "bootstrap/raw_results_structure")
bootstrap.ParTab(myBoot, idx = list(theta=1:13,omega=1:7,sigma=1))
}

