% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcoxf.r
\name{boxcoxf}
\alias{boxcoxf}
\title{Apply a Box-Cox Transformation to a Vector}
\usage{
boxcoxf(shape, distr)
}
\arguments{
\item{shape}{shape parameter}

\item{distr}{the distribution}
}
\value{
Box-Cox tranformation of \code{distr} with \code{shape}.
}
\description{
Applies a Box-Cox transformation to a vector.
}
\examples{
x = rnorm(1000, mean = 0, sd = 0.25)
x.boxcox.pos = boxcoxf(0.5,x)
x.boxcox.neg = boxcoxf(-3.5,x)
par(mfrow = c(1,3))
hist(x)
hist(x.boxcox.pos)
hist(x.boxcox.neg)
}
