% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfc.r
\name{cfc}
\alias{cfc}
\title{Create a vector, matrix, or data frame from clipboard object}
\usage{
cfc(obj.type = "data.frame", strings.as.factors = FALSE, row.names = NULL)
}
\arguments{
\item{obj.type}{either one of 'data.frame' (default), 'vector' or 'matrix'}

\item{strings.as.factors}{logical to treat strings as factors}

\item{row.names}{determining row.names}
}
\description{
Create a vector, matrix, or data frame from clipboard object
}
