% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conDataFun1.r
\name{conDataFun1}
\alias{conDataFun1}
\title{Summary: mean (SD)}
\usage{
conDataFun1(y, digits = 3, latex = FALSE, align.dot = FALSE, na.rm = FALSE)
}
\arguments{
\item{y}{numeric vector to summarize}

\item{digits}{number of significant digits}

\item{latex}{passed to \code{\link{formatted.signif}}}

\item{align.dot}{passed to \code{\link{formatted.signif}}}

\item{na.rm}{passed to \code{\link{mean}}, \code{\link{var}}}
}
\description{
Summary: mean (SD)
}
\note{
This function is primarily used for demographics tables
}
\examples{
x = rnorm(1000)
conDataFun1(x, 3)
}
\seealso{
\code{\link{conDataFun2}},  \code{\link{conDataFun3}},  \code{\link{catDataFun}}, \code{\link{tabStats}}, \code{\link{tabStats}}
}
