% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halflife.r
\name{ConvertPK.iv.Vk.1comp}
\alias{ConvertPK.iv.Vk.1comp}
\title{Calculates coefficients and exponents based on microconstants for 1-compartment IV popPK (post-hoc) parameter estimates}
\usage{
ConvertPK.iv.Vk.1comp(
  ds,
  subjVar = "ID",
  Dose = "DOSE",
  V1 = "V1",
  k10 = "K10",
  nsig = 3,
  ...
)
}
\arguments{
\item{ds}{data frame}

\item{subjVar}{name of the variable that represents the subject identifier}

\item{V1}{name of the variable that represents the central volume of distribution}

\item{k10}{name of the variable that represents the microconstant k10}

\item{nsig}{the number of significant digits}

\item{Dosename}{name of the variable that represents the dose}
}
\value{
returns a data.frame with coefficients, exponents and derived parameters
}
\description{
Calculates coefficients and exponents based on microconstants for 1-compartment IV popPK (post-hoc) parameter estimates
}
\examples{

X <- data

ConvertPK.iv.Vk.1comp(x)

}
\references{
Dennis Fisher & Steven Shafer NONMEM Workshop - Basic Concepts (2007)
}
\author{
Koen Jolling, Max Lagraauw
}
