% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp.examples.r
\name{example.xpose.VPC}
\alias{example.xpose.VPC}
\title{qPharmetra examples for plot, VPC and NONMEM dataset}
\usage{
example.xpose.VPC()
}
\value{
text (\code{example.xpose.VPC} and \code{example.CWRES.plot}) or a NONMEM data.frame (\code{example.NONMEM.dataset})
}
\description{
examples for qPharmetra style VPC and GOF plots and a NONMEM data set
}
\examples{
example.xpose.VPC()

example.CWRES.plot()

nmData = example.NONMEM.dataset()
nmData \%>\% group_by(DOSE) \%>\% summarise(nID=lunique(ID), nObs = length(DV))

pkpdData = example.pkpdData()
tbl_df(pkpdData)
}
