% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvpc_standard.r
\name{ggvpc_standard}
\alias{ggvpc_standard}
\title{VPC plot using ggplot2 (standard)}
\usage{
ggvpc_standard(
  vpc,
  PI = c(0.025, 0.975),
  area.col = PI.ci.med.arcol,
  linecol.obs.central = PI.real.med.col,
  linetype.obs.central = "solid",
  linetype.obs.outer = "dashed",
  linecol.obs.outer = "darkslategrey",
  linesize.obs = 0.5,
  alpha = 0.33,
  point.shape = 1,
  point.size = 1.5,
  point.col = "darkslategrey",
  yrange.stretch = c(0.9, 1.1),
  quiet = TRUE
)
}
\arguments{
\item{vpc}{output from \code{nm.read.vpc}}

\item{PI}{prediction interval (c(0.025,0.975) for 95\% CI)}

\item{area.col}{color of prediction polygon}

\item{linecol.obs.central}{color of observed central lines}

\item{linetype.obs.central}{line type of observed central lines}

\item{linetype.obs.outer}{line type of observed outer lines}

\item{linecol.obs.outer}{line color of observed outer lines}

\item{linesize.obs}{line width of observed data}

\item{alpha}{transparancy scalar (between 0 and 1)}

\item{point.shape}{numeric value for dot shape}

\item{point.size}{numeric value for dot size}

\item{point.col}{color of observed data dots}

\item{yrange.stretch}{vector of c(min,max) which will proportionally rescale the lower and upper limits of the Y axis}

\item{quiet}{whether to suppress process messages}
}
\value{
A ggplot object to be extended optionally
}
\description{
Plot basic vpc with 95\% PI and observed data. This call is to be extended with for stratification, and axes definition using the layering custom to ggplot2 objects. See the examples below to learn more.
}
\note{
Editing and stratification to be done by adding ggplot layer
}
\examples{
library(ggplot2)
vpc.all = nm.read.vpc(path = file.path(getOption("qpExampleDir"), "vpc_final_strt"))
ggvpc_standard(vpc.all)

ggvpc_standard(vpc.all) +
  labs(x="Time (h)", y="Concentration (ng/ml)")

## different Prediction Interval (PI), let's use 90\% instead of 95\%
ggvpc_standard(vpc.all, PI = c(0.05,0.95)) +
  labs(x="Time (h)", y="Concentration (ng/ml)")

## logging axes
ggvpc_standard(vpc.all) +
  labs(x="Time (h)", y="Concentration (ng/ml)") + scale_y_log10() + scale_x_log10()

#modify colors and transparency:
ggvpc_standard(vpc.all, area.col = indigo, alpha = 1, point.shape = 15) +
  labs(x="Time (h)", y="Concentration (ng/ml)") + scale_y_log10()

## dealing with stratification
myVPC = nm.read.vpc(path = file.path(getOption("qpExampleDir"), "vpc_base_strt"))

p = ggvpc_standard(vpc.all, yrange.stretch = c(1,1))
p = p +  labs(x="Time (h)", y="Concentration (ng/ml)")
p +  facet_wrap(~strata) + scale_y_log10()
}
\seealso{
\code{\link{nm.read.vpc}},  \code{\link{ggvpc_xpose}}
}
