% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nm.process.coveffects.r
\name{nm.process.coveffects}
\alias{.coveffect}
\alias{bootsamp}
\alias{covInfo}
\alias{coveff.plot}
\alias{eval.eqs}
\alias{make.eqs}
\alias{makeCovInfo}
\alias{nm.process.coveffects}
\alias{print.covInfo}
\title{Evaluate equations (as a list of calls) over the covariates involved.}
\usage{
nm.process.coveffects(eqs, covs.info, pars, xpose.df, omega, Nboot = 10000,
  pi.wid = 0.9)

coveff.plot(xpose.df, ppred.df, box, grouping)

make.eqs(...)

eval.eqs(.data, ...)

covInfo(cov.name, cov.vals, cov.cat = F, cov.center = median(cov.vals),
  cov.min = min(cov.vals), cov.max = max(cov.vals), cov.label = cov.name,
  cov.breaks = pretty(cov.vals))

print.covInfo(x)

makeCovInfo(df, cnames = names(df), cat.covs = "")

bootsamp(equ, .data, varcov, N = 10000)

.coveffect(equ, cov.info, ptab, varcov, Nboot = 10000, pi.wid = 0.9)
}
\arguments{
\item{eqs}{A list of equation calls}

\item{covs.info}{A list of covInfo objects to parameterize the subanalysis}

\item{xpose.df}{A dataframe containing post-hoc parameter values}

\item{omega}{The random effect covariance matrix from which to simulate ETA values}

\item{Nboot}{The number of bootstrap samples to run for each sample point}

\item{pi.wid}{The width of prediction intervals (e.g. 0.9 is a 90\% prediction interval)}

\item{box}{boolean for boxplots. It is set to TRUE if x axis is a categorical covariate.}

\item{grouping}{define the grouping level}

\item{...}{A series of named formulas}

\item{.data}{is data.frame, list, or environment}

\item{cov.name}{The name of the covariate}

\item{cov.vals}{A vector of values}

\item{cov.cat}{A boolean indicating if the covariate is categorical}

\item{cov.center}{The typical value of the covariate}

\item{cov.min}{The minimum value of the covariate}

\item{cov.max}{The maximum value of the covariate}

\item{cov.label}{The printed name of the covariate (e.g. seen on plot axes)}

\item{cov.breaks}{The points at which simulations and plots would be assessed}

\item{x}{A covInfo object}

\item{df}{A data frame}

\item{cnames}{A vector of column names to create covInfo objects for}

\item{cat.covs}{A vector of categorical column names.  Must be a subset of cnames}

\item{equ}{output from \code{make.eqs}}

\item{varcov}{variance-covariance matrix of OMEGA point estimates.}

\item{N}{sample size for simulation.}

\item{cov.info}{covariate center, min, max, breaks.}

\item{ptab}{TH/OM/SI point estimates as either a 1-row data frame or a flat list.}

\item{xpose.df}{output from \code{get.xpose.tables}}

\item{ppred}{prediction data frame}

\item{...}{is a list of calls or expressions}

\item{.data}{data frame to evaluate equations over.}

\item{equ}{output from \code{make.eqs}.}

\item{varcov}{variance-covariance matrix of OMEGA point estimates.}

\item{Nboot}{sample size for simulation.}

\item{pi.wid}{upper quantile of prediction interval. Defaults to 0.9.}
}
\value{
\code{nm.process.coveffects} A list including lists of: covariate effect plots, table of covariate effects,
table of post-hoc values -- for each covariate effect; the call list; the list of covInfo

\code{coveff.plot} A plot in a given environment.

\code{make.eqs} A list of call objects that can be used to evaluate the input equations
in a given environment.

.data is returned with the evaluated expressions added

\code{covInfo} A covInfo object

\code{print.covInfo} A covInfo object

\code{makeCovInfo} A list of covInfo objects

\code{bootsamp} Mean, median, variance of sample.

\code{.coveffect} A data.frame containing coveffect grid.
}
\description{
Covariate analysis is an essential element of population PK and PK-PD analysis. While there are useful tools available to analyse the impact of covariates, there is no tool available that visualizes or tabulates the impact of covariates on parameters they are affecting.

The main function \code{nm.process.coveffects} combines the input of helper functions to create a list of plots and tables detailing the covariate relationships as declared. The combined set of equations define the covariate relationships by evaluating declared relationships between parameters and covariates across all covariate relationships defined and provides tabulation as well as detailed graphical displays.

The function is set up such that estimated entities such as THETAs and OMEGAs can be integrated easily by using function from \code{qpToolkit}.

\code{coveff.plot} prints plot (a) object(s)

\code{make.eqs} makes a list of calls from its named arguments
input formulas.

\code{eval.eqs} evaluates a list of calls in an environment, list, or data.frame.

\code{covInfo} constructs a covInfo (covariate information) object from a vector.

\code{print.covInfo} prints a covInfo object.

\code{makeCovInfo} takes a data frame and creates covInfo objects based on its columns.

\code{bootsamp} samples an equation with random normals pulled from varcov, other variables from supplied env.

\code{.coveffect} samples an equation with random normals pulled from varcov, other variables from supplied arguments.
}
\section{Functions}{
\itemize{
\item \code{coveff.plot}: 

\item \code{make.eqs}: 

\item \code{eval.eqs}: 

\item \code{covInfo}: 

\item \code{print.covInfo}: 

\item \code{makeCovInfo}: 

\item \code{bootsamp}: 

\item \code{.coveffect}: 
}}
\note{
This function is used with \code{nm.process.coveffects}
}
\examples{
eqs=make.eqs(CL=THETA1*(1 + (GNDR==0)*THETA7)*(AGE/33.72)^THETA6*exp(ETA1)
            ,V1=THETA2*(1 + (GNDR==0)*THETA9)*(AGE/33.72)^THETA8*exp(ETA2)
            )
eqs$label=list(CL="Clearance (L/h)", V1="Volume of Central Compartment (L)")
}

