% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.process.coveffects.r
\name{nm.process.coveffects}
\alias{nm.process.coveffects}
\title{Evaluate equations (as a list of calls) over the covariates involved.}
\usage{
nm.process.coveffects(eqs, covs.info, pars, xpose.df, omega,
  Nboot = 10000, pi.wid = 0.9)
}
\arguments{
\item{eqs}{A list of calls}

\item{covs.info}{A list of covInfo objects to parameterize the subanalysis}

\item{pars}{A list or single-row data.frame of the parameter estimates (THETA, OMEGA)}

\item{xpose.df}{A dataframe containing post-hoc parameter values}

\item{omega}{The random effect covariance matrix from which to simulate ETA values}

\item{Nboot}{The number of bootstrap samples to run for each sample point}

\item{pi.wid}{The width of prediction intervals (e.g. 0.9 is a 90\% prediction interval)}
}
\value{
A list including lists of: covariate effect plots, table of covariate effects,
table of post-hoc values -- for each covariate effect; the call list; the list of covInfo
}
\description{
A list of equations that define the covariate relationships is...
}
