% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.process.scm.r, R/nm.process.scm_old.r
\name{nm.process.scm}
\alias{nm.process.scm}
\title{Parsing SCM output}
\usage{
nm.process.scm(path, ...)

nm.process.scm(path, ...)
}
\arguments{
\item{path}{full directory path where scmlog files reside}

\item{...}{ignored}
}
\value{
data.frame

data.frame
}
\description{
Read and process SCM output file such that the output can be directly copied into Microsoft products and LaTeX typesetting.

Read and process SCM output file such that the output can be directly copied into Microsoft products and LaTeX typesetting.
}
\examples{
library(magrittr)
library(dplyr)
library(knitr)
path = file.path(getOption("qpExampleDir"),"scm_example2")
nm.process.scm(path)
  nm.process.scm \%>\% 
  filter(chosen == 1) \%>\%                      # summary
  select(step, model, direction, pvalue) \%>\%
  formalize \%>\%                                # labels as column names
  decode \%>\%                                   # substitute guide values 
  kable
 # model:
 path \%>\% nm.process.scm \%>\% filter(chosen == 1) \%>\%
 select(model)
library(encode)
library(magrittr)
library(dplyr)
library(knitr)
path = file.path(getOption("qpExampleDir"),"scm_example2")
nm.process.scm(path)
path \%>\%
  nm.process.scm \%>\% 
  filter(chosen == 1) \%>\%                      # summary
  select(step, model, direction, pvalue) \%>\%
  formalize \%>\%                                # labels as column names
  decode \%>\%                                   # substitute guide values 
  kable
 # model:
 path \%>\% nm.process.scm \%>\% filter(chosen == 1) \%>\%
 select(model)
}
