% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.process.scm.r
\name{nm.process.scm}
\alias{nm.process.scm}
\title{Parsing SCM output}
\usage{
nm.process.scm(path, ...)
}
\arguments{
\item{path}{full directory path where scmlog files reside}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Read and process SCM output file such that the output can be directly copied into Microsoft products and LaTeX typesetting.
}
\examples{
library(encode)
library(magrittr)
library(dplyr)
library(knitr)
path = file.path(getOption("qpExampleDir"),"scm_example2")
nm.process.scm(path)
path \%>\%
  nm.process.scm \%>\% 
  filter(chosen == 1) \%>\%                      # summary
  select(step, model, direction, pvalue) \%>\%
  formalize \%>\%                                # labels as colum names
  decode \%>\%                                   # substitute guide values 
  kable
 # model:
 path \%>\% nm.process.scm \%>\% filter(chosen == 1) \%>\%
 select(model)
}
