% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nm.process.scm.r
\name{nm.process.scm}
\alias{nm.process.scm}
\title{Parsing SCM output}
\usage{
nm.process.scm(path)
}
\arguments{
\item{path}{full directory path where scmlog files reside}
}
\value{
A character vector with processed SCM output, ready for inclusion into a (LaTeX) report
}
\description{
Read and process SCM output file such that the output can be directly copied into Microsoft products and LaTeX typesetting.
}
\examples{
## the code below works if a number of scmlog.txt files are in folder NONMEM/scm files
myPath = file.path(getOption("qpExampleDir"),"scm_example2")
scm = nm.process.scm(myPath)
names(scm)

scm$summary

# populate an appendix with scm output
print(xtable(scm$scm.latex
   , caption = "Full Stepwise Covariate Building output"
   , label =  'Tab:scmfull')
      , sanitize.text.function = function(x)x
      , include.rownames = F
      , size = "small"
      , caption.placement ="top"
      , booktabs = TRUE
      , tabular.environment = "longtable"
      , floating = FALSE
      #  , only.contents = T
)
}

