% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.cwres.r
\name{panel.cwres}
\alias{panel.cwres}
\title{Lattice panel function for CWRES plots}
\usage{
panel.cwres(
  x,
  y,
  ...,
  col = obs.color,
  col.line = abline.color,
  pch = 1,
  col.loess = smooth.color,
  family = "gaussian"
)
}
\arguments{
\item{x, y}{numeric vectors x and y}

\item{...}{any other arguments to be passed on to the lattice call}

\item{col}{symbol color (defaults to obs.color)}

\item{col.line}{color of abline}

\item{pch}{symbol}

\item{col.loess}{color of smooth line}

\item{family}{passed to \code{loess.smooth} (by way of \code{\link[lattice]{panel.loess}})}
}
\value{
Lattice panel output (invisible)
}
\description{
Panel function to plot CWRES vs any covariate
}
\examples{
library(lattice)
library(reshape2)
 out = get.xpose.tables("example2", getOption("qpExampleDir"))
 trellis.strip.color()
xyplot(CWRES ~ value | variable
   , data = subset(reshape2::melt(out, measure.vars = c('PRED','TIME')),EVID == 0)
   , aspect = 1
   , scales = list(x = list(relation = "free"))
   , panel = panel.cwres,
   family = 'symmetric'
   )
}
