% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.nlme.vpc.obs.r
\name{panel.nlme.vpc.obs}
\alias{panel.nlme.vpc.obs}
\title{Lattice function for observed components of an nlme VPCs}
\usage{
panel.nlme.vpc.obs(
  x,
  y,
  logY,
  showPredAs,
  showObsDots,
  showObsLines,
  col.scheme,
  obscex.dot,
  obspch.dot,
  ...
)
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}

\item{logY}{unused}

\item{showPredAs}{unused}

\item{showObsDots}{whether to show observations as dots}

\item{showObsLines}{whether to show observations as lines}

\item{col.scheme}{list with elements: [obs: [dot, line]]}

\item{obscex.dot}{cex for observations}

\item{obspch.dot}{pch for observations}

\item{\dots}{passed to panel.xyplot()}
}
\value{
Nothing -> internal to a lattice call
}
\description{
Panel function for observed lines / colored polygons of a vpc plot of an nlme object
}
\note{
Not to be used as standalone. Will be be called by \code{nlme.vpcplot}
}
\seealso{
\code{\link{nlme.vpcplot}}
}
