% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/panel.nonmem.vpc.r
\name{panel.nonmem.vpc}
\alias{panel.nonmem.vpc}
\title{Lattice function for predicted components of a NONMEM VPC}
\usage{
panel.nonmem.vpc(x, y, OBS, vpc, subscripts, logY, logX, showPredAs,
  showObsDots, showObsLines, xCov, col.scheme, obscex.dot, obspch.dot, lineType,
  central.lwd, ...)
}
\arguments{
\item{logY}{logical for logging Y coordinate}

\item{logX}{logical for logging X coordinate}

\item{other}{a collection of parameters are available to modify the layout of the VPC}
}
\value{
Lattice panel output (invisible)
}
\description{
Panel function for predicted and observed lines / colored polygons of a vpc plot of an NONMEM / PsN VPC object
}
\note{
Not to be used as standalone. Will be be called by \code{nm.vpcplot}. Furthermore, the \code{nm.vpcplot} routine will go deprecated soon, thus it is not adviced to use this function going forward. Instead use xpose.VPC.
}
\seealso{
\code{\link{nm.vpcplot}}
}

