% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.process.coveffects.r
\name{plot.covEffects}
\alias{plot.covEffects}
\title{Plot a covariate effects object}
\usage{
\method{plot}{covEffects}(x, y = names(x$plot), ...)
}
\arguments{
\item{x}{a covEffects object from nm.process.coveffects}

\item{y}{names of the plots to print (defaults to all)}

\item{\dots}{ignored}
}
\description{
Plot a covariate effects object
}
