% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nm.process.coveffects.r
\name{plot.covEffects}
\alias{plot.covEffects}
\title{Plot a cov effects object}
\usage{
\method{plot}{covEffects}(obj, idx = names(obj$plot))
}
\arguments{
\item{obj}{output from nm.process.coveffects}

\item{idx}{names or list indices of the plots to print}
}
\value{
\code{plot.covEffects} A plot in a given environment.
}
\description{
\code{plot.covEffects} prints a plot object
}
\note{
This function is used with \code{nm.process.coveffects}
}
\seealso{
\code{\link{nm.process.coveffects}}
}

