% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.process.coveffects.r
\name{print.covInfo}
\alias{print.covInfo}
\title{Print a covInfo S3 object}
\usage{
\method{print}{covInfo}(x, ...)
}
\arguments{
\item{x}{A covInfo object}

\item{\dots}{ignored}
}
\value{
A covInfo object
}
\description{
Prints a covInfo object.
}
\details{
Prints a covInfo object.
}
\note{
This function is used with \code{nm.process.coveffects}
}
\seealso{
Other coveffects: \code{\link{covInfo}},
  \code{\link{eval.eqs}}, \code{\link{make.eqs}}
}
\concept{coveffects}
