% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qP.system.R
\name{qP.welcome}
\alias{qP.welcome}
\title{Print a welcome message}
\usage{
qP.welcome(name = "")
}
\arguments{
\item{name}{length one character: who is being addressed}
}
\value{
A welcoming message
}
\description{
Print a welcome message
}
\keyword{internal}
