% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpToolkit.R
\docType{package}
\name{qpToolkit}
\alias{qpToolkit}
\alias{qpToolkit-package}
\title{qpToolkit: Makes qPharmetra R functions easily available to users}
\description{
The qpToolkit package provides different categories of important functions:
\itemize{
\item{R modeling functions}
\item{NONMEM post-processing}
\item{Data tables}
\item{R qualification}}
}
\section{qpToolkit functions}{

The qpToolkit functions ...
}

