% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabSummarize.r
\name{tabSummarize}
\alias{tabSummarize}
\title{Tabulate a summary}
\usage{
tabSummarize(
  formula,
  data,
  digits = 3,
  extra.blank.line = TRUE,
  digits.categorical = 1,
  na.rm = FALSE
)
}
\arguments{
\item{formula}{formula specifying what to summarize by what}

\item{data}{data to summarize}

\item{digits}{number of significant digits in output}

\item{extra.blank.line}{a logical indicating if an empty line needs to be inserted between two summaries to improve layout and legibility}

\item{digits.categorical}{passed to \code{\link{tabStats}}}

\item{na.rm}{passed to \code{\link{tabStats}}}
}
\value{
Table with summarized data, of class \code{demoTable}
}
\description{
Tabulate a summary
}
\note{
This function is primarily used for demographics tables
}
\examples{
options(width = 150)
pkpdData = example.pkpdData()
ok = duplicated(pkpdData$id) == FALSE
tabSummarize(formula = dose ~ race + wt + bmi + sex, data = pkpdData[ok, ], digits = 3)
pkpdData$sex <- factor(pkpdData$sex, levels = c('M','F'))
tabSummarize(formula = dose ~ race + wt + bmi + sex, data = pkpdData[ok, ], digits = 3)
# check the categorical summary for race
round(table(pkpdData$race[ok], pkpdData$dose[ok]) / apply(table(pkpdData$race[ok],
pkpdData$dose[ok]), 2, sum) * 100)
# OK
myFormula =  dose ~ race + wt + bmi + sex
format(
 tabSummarize(formula = myFormula, data = pkpdData[ok, ], digits = 3)
  , formula = myFormula
)
}
\seealso{
\code{\link{conDataFun1}},  \code{\link{conDataFun2}},  \code{\link{conDataFun3}},  \code{\link{catDataFun}},  \code{\link{tabStats}}
}
